% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaIndex.R
\docType{methods}
\name{alphaindex}
\alias{alphaindex}
\alias{alphaindex,matrix-method}
\alias{alphaindex,matrix}
\alias{alphaindex,data.frame-method}
\alias{alphaindex,data.frame}
\alias{alphaindex,integer-method}
\alias{alphaindex,integer}
\alias{alphaindex,phyloseq-method}
\alias{alphaindex,phyloseq}
\title{alpha index}
\usage{
alphaindex(obj, ...)

\S4method{alphaindex}{matrix}(obj, mindepth, sampleda, ...)

\S4method{alphaindex}{data.frame}(obj, ...)

\S4method{alphaindex}{integer}(obj, ...)

\S4method{alphaindex}{phyloseq}(obj, ...)
}
\arguments{
\item{obj}{object, data.frame of (nrow sample * ncol taxonomy(feature)) 
or phyloseq.}

\item{...}{additional arguments.}

\item{mindepth}{numeric, Subsample size for rarefying community.}

\item{sampleda}{data.frame,sample information, row sample * column factors.}
}
\value{
data.frame contained alpha Index.
}
\description{
caculate the alpha index (Obseve,Chao1,Shannon,Simpson) of sample
with \code{\link[vegan]{diversity}}
}
\examples{
library(tidyverse)
otudafile <- system.file("extdata", "otu_tax_table.txt", 
                        package="MicrobiotaProcess")
otuda <- read.table(otudafile, sep="\\t", 
             header=TRUE, row.names=1, 
             check.names=FALSE, skip=1, comment.char="")
otuda <- otuda[sapply(otuda, is.numeric)] \%>\% t() \%>\% 
          data.frame(check.names=FALSE)
set.seed(1024)
alphatab <- alphaindex(otuda)
head(as.data.frame(alphatab))
data(test_otu_data)
class(test_otu_data)
set.seed(1024)
alphatab2 <- alphaindex(test_otu_data)
head(as.data.frame(alphatab2))
}
\author{
ShuangbinXu
}
