% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{alphasample-class}
\alias{alphasample-class}
\title{alphasample class}
\description{
alphasample class
}
\section{Slots}{

\describe{
\item{\code{alpha}}{data.frame contained alpha metrics of samples}

\item{\code{sampleda}}{associated sample information}
}}

