% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-method.R
\name{as.data.frame.diffAnalysisClass}
\alias{as.data.frame.diffAnalysisClass}
\title{get the table of diffAnalysisClass}
\usage{
\method{as.data.frame}{diffAnalysisClass}(x, ...)
}
\arguments{
\item{x}{object, diffAnalysisClass}

\item{..., }{additional parameters}
}
\value{
a data.frame contained results of diffAnalysis
}
\description{
get the table of diffAnalysisClass
}
\examples{
data(kostic2012crc)
kostic2012crc
head(phyloseq::sample_data(kostic2012crc),3)
kostic2012crc <- phyloseq::rarefy_even_depth(kostic2012crc,rngseed=1024)
table(phyloseq::sample_data(kostic2012crc)$DIAGNOSIS)
set.seed(1024)
diffres <- diffAnalysis(kostic2012crc, class="DIAGNOSIS",
                        mlfun="lda", filtermod="fdr",
                        firstcomfun = "kruskal.test",
                        firstalpha=0.05, strictmod=TRUE,
                        secondcomfun = "wilcox.test",
                        subclmin=3, subclwilc=TRUE,
                        secondalpha=0.01, lda=3)
restab <- as.data.frame(diffres)
head(restab)
}
