% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-as-MPSE.R
\docType{methods}
\name{as.MPSE}
\alias{as.MPSE}
\alias{as.MPSE,tbl_mpse-method}
\alias{as.MPSE,tbl_mpse}
\alias{as.MPSE,phyloseq-method}
\alias{as.MPSE,phyloseq}
\alias{as.MPSE,grouped_df_mpse-method}
\alias{as.MPSE,grouped_df_mpse}
\alias{as.MPSE,TreeSummarizedExperiment-method}
\alias{as.MPSE,TreeSummarizedExperiment}
\title{as.MPSE method}
\usage{
as.MPSE(.data, ...)

\S4method{as.MPSE}{tbl_mpse}(.data, ...)

\S4method{as.MPSE}{phyloseq}(.data, ...)

\S4method{as.MPSE}{grouped_df_mpse}(.data, ...)

\S4method{as.MPSE}{TreeSummarizedExperiment}(.data, ...)
}
\arguments{
\item{.data}{tbl_mpse or phyloseq or TreeSummarizedExperiment object}

\item{...}{additional parameters.}
}
\value{
MPSE object
}
\description{
as.MPSE method
}
\examples{
data(test_otu_data)
test_otu_data \%>\% as.MPSE -> mpse
mpse
}
\author{
Shuangbin Xu
}
