% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-as.treedata.R
\name{as.treedata.taxonomyTable}
\alias{as.treedata.taxonomyTable}
\title{as.treedata}
\usage{
\method{as.treedata}{taxonomyTable}(tree, include.rownames = FALSE, ...)
}
\arguments{
\item{tree}{object, This is for taxonomyTable class, 
so it should be a taxonomyTable.}

\item{include.rownames}{logical, whether to set the rownames of 
taxonomyTable to tip labels, default is FALSE.}

\item{...}{additional parameters.}
}
\description{
convert taxonomyTable to treedata
}
\examples{
\dontrun{
  data(test_otu_data)
  tree <- as.treedata(phyloseq::tax_table(test_otu_data), include.rownames = TRUE)
}
}
