% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{clustplotClass-class}
\alias{clustplotClass-class}
\title{clustplotClass class}
\description{
clustplotClass class
}
\section{Slots}{

\describe{
\item{\code{hclustphylo}}{phylo object (convert hclust to phylo).}

\item{\code{sampleda}}{assocaited sample information.}

\item{\code{distmethod}}{character the method of dist.}
}}

