% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountOrRatios.R
\name{CountOrRatios}
\alias{CountOrRatios}
\title{caculate the count or relative abundance of replicate element with a speficify column}
\usage{
CountOrRatios(da, featurelist = NULL, countmode = FALSE,
  percentmode = FALSE, multiplenum = 1, rownamekeep = FALSE)
}
\arguments{
\item{da}{dataframe; a dataframe contained one character column and others is numeric 
if featurelist is NULL, else a numeirc dataframe.}

\item{featurelist}{dataframe; a dataframe contained one chatacter column, default is NULL.}

\item{countmode}{boolean; whether return the count results (FALSE),or relative abundance (TRUE).}

\item{percentmode}{boolean; whether return the 100% percent (FALSE),or (0.1) percent(TRUE).}

\item{multiplenum}{numeric; the multiple you want to increase,default is 1.}

\item{rownamekeep}{boolean; whether you return a dataframe contained the rownames,default is FALSE.}
}
\description{
Caculate the count or relative abundance of replicate element with a speficify columns
}
\author{
Shuangbin Xu
}
