% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldata.R
\docType{data}
\name{data-hmp_aerobiosis_small}
\alias{data-hmp_aerobiosis_small}
\alias{hmp_aerobiosis_small}
\title{(Data) Small subset of the HMP 16S dataset}
\description{
Contained three datasets, featureda, sampleda, taxda
featureda contained 55 samples (nrow) and 1091 features (ncol)
sampleda contained 55 samples from 6 body sites of 10 subjects.
taxda contained 699 taxonomy by 6 rank. This datasets were built 
from the LEfSe.\url{http://huttenhower.sph.harvard.edu/webfm_send/129}
}
\examples{
data(hmp_aerobiosis_small)
}
\keyword{data}
