% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldata.R
\docType{data}
\name{data-kostic2012crc}
\alias{data-kostic2012crc}
\alias{kostic2012crc}
\title{(Data) Genomic analysis identifies association of Fusobacterium 
              with colorectal carcinoma (2012)}
\description{
This dataset was from the a study on colorectal cancer, 
publised in Genome Research (2012). This dataset had been removed samples 
with less than 500 reads, contained 91 Control and 86 Tumors. And It is 
belong to MPSE class, contained otu_table and sample_data.
}
\examples{
data(kostic2012crc)
}
\keyword{data}
