% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alldata.R
\docType{data}
\name{data-test_otu_data}
\alias{data-test_otu_data}
\alias{test_otu_data}
\title{(Data) simulated dataset.}
\description{
This dataset was simulated. And it also was phyloseq class, 
contained otu_table and sample_data
}
\examples{
data(test_otu_data)
}
\keyword{data}
