% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-method.R
\name{diffAnalysis}
\alias{diffAnalysis}
\alias{diffAnalysis.data.frame}
\title{Differential expression analysis based on kruskal.test, generalized fold change and wilcox.test}
\usage{
diffAnalysis(x, ...)

\method{diffAnalysis}{data.frame}(data, sampleda, class, subclass = NULL,
  firstcomfun = "kruskal.test", padjust = "fdr",
  filtermod = "pvalue", firstalpha = 0.05, strictmod = TRUE,
  fcfun = "generalizedFC", secondcomfun = "wilcox.test", clmin = 5,
  clwilc = TRUE, submin = 5, subclwilc = TRUE, secondalpha = 0.05,
  ...)
}
\description{
Differential expression analysis based on kruskal.test, generalized fold change and wilcox.test
}
\author{
Shuangbin Xu
}
