% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-binary_offspring.R
\name{extract_binary_offspring}
\alias{extract_binary_offspring}
\title{extract the binary offspring of the specified internal nodes}
\usage{
extract_binary_offspring(.data, .node, type = "tips", ...)
}
\arguments{
\item{.data}{phylo or treedata object}

\item{.node}{the internal nodes}

\item{type}{the type of binary offspring, options are 'tips' (default),
'all', 'internal'.}

\item{...}{additional parameter, meaningless now.}
}
\description{
extract the binary offspring of the specified internal nodes
}
