% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_ord_ellipse.R
\name{geom_ord_ellipse}
\alias{geom_ord_ellipse}
\title{add confidence ellipse to ordinary plot}
\usage{
geom_ord_ellipse(
  data = NULL,
  mapping = NULL,
  ellipse_pro = 0.9,
  alpha = 0.3,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{data, }{default is NULL.}

\item{mapping, }{aes mapping, default is NULL.}

\item{ellipse_pro}{numeric, confidence value for the ellipse, default is 0.9}

\item{alpha}{numeric, alpha of ellipse, default is 0.2}

\item{show.legend, }{default is NA.}

\item{inherit.aes, }{default is TRUE}

\item{...}{additional parameters}
}
\value{
ggplot layer
}
\description{
add confidence ellipse to ordinary plot
}
\references{
\url{http://lchblogs.netlify.com/post/2017-12-22-r-addconfellipselda/}
}
\author{
Guangchuang Yu
}
\keyword{internal}
