% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-get-pcoa.R, R/plot-ggordpoint.R
\name{get_coord.pcoa}
\alias{get_coord.pcoa}
\alias{get_coord}
\alias{get_coord.prcomp}
\title{get ordination coordinates.}
\usage{
\method{get_coord}{pcoa}(obj, pc)

get_coord(obj, pc)

\method{get_coord}{prcomp}(obj, pc)
}
\arguments{
\item{obj}{object,prcomp class or pcoa class}

\item{pc}{integer vector, the component index.}
}
\value{
ordplotClass object.
}
\description{
get ordination coordinates.
}
\examples{
require(graphics)
data(USArrests)
pcares <- prcomp(USArrests, scale = TRUE)
coordtab <- get_coord(pcares,pc=c(1, 2))
coordtab2 <- get_coord(pcares, pc=c(2, 3))
}
