% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-utilities.R
\name{get_pvalue}
\alias{get_pvalue}
\alias{get_pvalue.htest}
\alias{get_pvalue.lme}
\alias{get_pvalue.negbin}
\alias{get_pvalue.ScalarIndependenceTest}
\alias{get_pvalue.QuadTypeIndependenceTest}
\alias{get_pvalue.lm}
\alias{get_pvalue.glm}
\title{Methods for computation of the p-value}
\usage{
get_pvalue(obj)

\method{get_pvalue}{htest}(obj)

\method{get_pvalue}{lme}(obj)

\method{get_pvalue}{negbin}(obj)

\method{get_pvalue}{ScalarIndependenceTest}(obj)

\method{get_pvalue}{QuadTypeIndependenceTest}(obj)

\method{get_pvalue}{lm}(obj)

\method{get_pvalue}{glm}(obj)
}
\arguments{
\item{obj}{object, such as htest, lm, negbin
ScalarIndependenceTest class.}
}
\value{
pvalue.
}
\description{
Methods for computation of the p-value
}
\examples{
library(nlme)
lmeres <- lme(distance ~ Sex,data=Orthodont)
pvalue <- get_pvalue(lmeres)
}
\author{
Shuangbin Xu
}
