% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{get_taxadf}
\alias{get_taxadf}
\alias{get_taxadf.phyloseq}
\alias{get_taxadf.default}
\title{get the data of specified taxonomy}
\usage{
get_taxadf(obj, ...)

\method{get_taxadf}{phyloseq}(obj, taxlevel = 2, ...)

\method{get_taxadf}{default}(obj, taxda, taxa_are_rows, taxlevel, sampleda = NULL, ...)
}
\arguments{
\item{obj}{phyloseq, phyloseq class or data.frame
the shape of data.frame (nrow sample * column feature
taxa_are_rows set FALSE, nrow feature * ncol sample, 
taxa_are_rows set TRUE).}

\item{...}{additional parameters, see also 
\code{\link[MicrobiotaProcess]{CountOrRatios}}}

\item{taxlevel}{character, the column names of taxda that you want to get.
when the input is phyloseq class, you can use 1 to 7.}

\item{taxda}{data.frame, the classifies of feature contained 
in obj(data.frame).}

\item{taxa_are_rows}{logical, if the column of data.frame
are features, it should be set FALSE.}

\item{sampleda}{data.frame, the sample information.}
}
\value{
phyloseq class contained tax data.frame and sample information.
}
\description{
get the data of specified taxonomy
}
\examples{
library(ggplot2)
data(test_otu_data)
phytax <- get_taxadf(test_otu_data, taxlevel=2)
phytax
head(phyloseq::otu_table(phytax))
phybar <- ggbartax(phytax) + 
         xlab(NULL) + ylab("relative abundance (\%)")
}
\author{
Shuangbin Xu
}
