% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ordination.R, R/get_pca.R
\name{get_varct.pcoa}
\alias{get_varct.pcoa}
\alias{get_varct}
\alias{get_varct.prcomp}
\alias{get_varct.pcasample}
\title{get the contribution of variables}
\usage{
\method{get_varct}{pcoa}(obj, ...)

get_varct(obj, ...)

\method{get_varct}{prcomp}(obj, ...)

\method{get_varct}{pcasample}(obj, ...)
}
\arguments{
\item{obj}{prcomp class or pcasample class}

\item{...}{additional parameters.}
}
\value{
the VarContrib class, contained the 
contribution and coordinate of features.
}
\description{
get the contribution of variables
}
\examples{
library(phyloseq)
data(GlobalPatterns)
subGlobal <- subset_samples(GlobalPatterns,
         SampleType \%in\% c("Feces", "Mock", "Ocean", "Skin"))
pcares <- get_pca(subGlobal, method="hellinger") 
varres <- get_varct(pcares)
}
