% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getclust.R
\name{getclust}
\alias{getclust}
\alias{getclust.dist}
\alias{getclust.default}
\alias{getclust.phyloseq}
\title{Hierarchical cluster analysis for the samples}
\usage{
getclust(obj, ...)

\method{getclust}{dist}(distobj, distmethod, sampleda = NULL,
  hclustmethod = "average", ...)

\method{getclust}{default}(data, distmethod = "euclidean",
  taxa_are_rows = FALSE, sampleda = NULL, tree = NULL,
  method = "hellinger", hclustmethod = "average", ...)

\method{getclust}{phyloseq}(phyloseqobj, distmethod = "euclidean",
  method = "hellinger", hclustmethod = "average", ...)
}
\arguments{
\item{obj}{R object,dist data.frame or phyloseq class}

\item{distmethod}{character, the method of dist, when the obj is data.frame 
or phyloseq default is "euclidean". see also \code{\link[MicrobiotaProcess]{getdist}}.}

\item{sampleda}{data.frame, nrow sample * ncol factor. default is NULL.}

\item{hclustmethod}{character, the method of hierarchical cluster, default is average.}
}
\description{
Hierarchical cluster analysis for the samples
}
\author{
ShuangbinXu
}
