% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getordination.R, R/getpca.R
\name{getcoord.pcoa}
\alias{getcoord.pcoa}
\alias{getcoord}
\alias{getcoord.prcomp}
\title{get ordination coordinates.}
\usage{
\method{getcoord}{pcoa}(obj, pc)

getcoord(obj, pc)

\method{getcoord}{prcomp}(obj, pc)
}
\arguments{
\item{obj}{object,prcomp class or pcoa class}

\item{pc}{integer vector, the component index.}
}
\value{
ordplotClass object.
}
\description{
get ordination coordinates.

get ordination coordinates
}
\examples{
require(graphics)
data(USArrests)
pcares <- prcomp(USArrests, scale = TRUE)
coordtab <- getcoord(pcares,pc=c(1, 2))
coordtab2 <- getcoord(pcares, pc=c(2, 3))
}
