% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdiffAnalysis.R
\name{getMeanMedian}
\alias{getMeanMedian}
\title{get the mean and median of specific feature.}
\usage{
getMeanMedian(datameta, feature, subclass)
}
\arguments{
\item{datameta}{data.frame, nrow sample * ncol feature + factor.}

\item{feature}{character vector, the feature contained in datameta.}

\item{subclass}{character, factor name.}
}
\description{
get the mean and median of specific feature.
}
\author{
Shuangbin Xu
}
