% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpca.R
\name{getpca.default}
\alias{getpca.default}
\title{getpca}
\usage{
\method{getpca}{default}(data, sampleda = NULL, method = "log", ...)
}
\arguments{
\item{data}{data.frame, nrow sample * ncol feature}

\item{sampleda}{data.frame, nrow sample * ncol factors.}

\item{method}{character, the standardization methods for 
community ecologists. see \code{\link[vegan]{decostand}}}
}
\description{
getpca
}
