% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpca.R
\name{getpca}
\alias{getpca}
\alias{getpca.default}
\alias{getpca.phyloseq}
\title{Performs a principal components analysis}
\usage{
getpca(obj, ...)

\method{getpca}{default}(data, sampleda = NULL, method = "log", ...)

\method{getpca}{phyloseq}(obj, ...)
}
\arguments{
\item{obj}{data.frame or phyloseq class,shape of data.frame is nrow sample * ncol feature}

\item{...}{additional parameters, see\code{\link[stat]{prcomp}}.}

\item{sampleda}{data.frame, nrow sample * ncol factors.}

\item{method}{character, the standardization methods for
community ecologists. see \code{\link[vegan]{decostand}}.}
}
\value{
pcasample class, contained prcomp class and sample information.
}
\description{
Performs a principal components analysis
}
