% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getordination.R
\name{getpcoa}
\alias{getpcoa}
\alias{getpcoa.default}
\alias{getpcoa.dist}
\alias{getpcoa.phyloseq}
\title{performs principal coordinate analysis (PCoA)}
\usage{
getpcoa(obj, ...)

\method{getpca}{default}(data, distmethod = "euclidean",
  taxa_are_rows = FALSE, sampleda = NULL, tree = NULL,
  type = "sample", method = "hellinger", ...)

\method{getpcoa}{dist}(obj, distmethod, data = NULL, sampleda = NULL,
  method = "hellinger", ...)

\method{getpcoa}{phyloseq}(obj, distmethod = "euclidean", ...)
}
\arguments{
\item{obj}{phyloseq, the phyloseq class or dist class.}

\item{..., }{additional parameter, see \code{\link[MicrobiotaProcess]{getdist}}.}

\item{data}{data.frame, numeric data.frame nrow sample * ncol features.}

\item{taxa_are_rows}{logical, if feature of data is column, it should be set FALSE.}

\item{sampleda}{data.frame, nrow sample * ncol factor, default is NULL.}

\item{tree}{phylo, the phylo class, default is NULL, when use unifrac method, it should be
required.}

\item{type}{character, default is sample.}

\item{method}{character, the standardization method for community ecologists, default is hellinger,
if the data has be normlized, it shound be set NULL.}
}
\description{
performs principal coordinate analysis (PCoA)
}
\author{
Shuangbin Xu
}
