% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-utilities.R
\name{getpvalue}
\alias{getpvalue}
\alias{getpvalue.htest}
\alias{getpvalue.lme}
\alias{getpvalue.negbin}
\alias{getpvalue.ScalarIndependenceTest}
\alias{getpvalue.QuadTypeIndependenceTest}
\alias{getpvalue.lm}
\alias{getpvalue.glm}
\title{Methods for computation of the p-value}
\usage{
getpvalue(obj)

\method{getpvalue}{htest}(obj)

\method{getpvalue}{lme}(obj)

\method{getpvalue}{negbin}(obj)

\method{getpvalue}{ScalarIndependenceTest}(obj)

\method{getpvalue}{QuadTypeIndependenceTest}(obj)

\method{getpvalue}{lm}(obj)

\method{getpvalue}{glm}(obj)
}
\arguments{
\item{obj}{object, such as htest, lm, negbin
ScalarIndependenceTest class.}
}
\value{
pvalue.
}
\description{
Methods for computation of the p-value
}
\examples{
library(nlme)
lmeres <- lme(distance ~ Sex,data=Orthodont)
pvalue <- getpvalue(lmeres)
}
\author{
Shuangbin Xu
}
