% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{gettaxdf.default}
\alias{gettaxdf.default}
\title{gettaxdf}
\usage{
\method{gettaxdf}{default}(otuda, taxda, taxlevel, sampleda = NULL, ...)
}
\arguments{
\item{otuda}{data.frame, otu table.}

\item{taxda}{data.frame, the dataframe of taxonomy.}

\item{taxlevel}{character, the column names of taxda that you want to get.}

\item{sampleda}{data.frame, the sample information.}
}
\description{
gettaxdf
}
