% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{gettaxdf.phyloseq}
\alias{gettaxdf.phyloseq}
\title{gettaxdf}
\usage{
\method{gettaxdf}{phyloseq}(obj, taxlevel = "Phylum", ...)
}
\arguments{
\item{obj}{the phyloseq object.}

\item{taxlevel}{character, default is "Phylum", options is "Kingdom",
"Phylum", "Class", "Order", "Family", "Genus", "Species".}

\item{...}{additional parameters.}
}
\description{
gettaxdf
}
