% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getordination.R, R/getpca.R
\name{getvarct.pcoa}
\alias{getvarct.pcoa}
\alias{getvarct}
\alias{getvarct.prcomp}
\alias{getvarct.pcasample}
\title{get the contribution of variables}
\usage{
\method{getvarct}{pcoa}(obj)

getvarct(obj, ...)

\method{getvarct}{prcomp}(obj)

\method{getvarct}{pcasample}(obj)
}
\arguments{
\item{obj}{prcomp class or pcasample class}
}
\description{
get the contribution of variables
}
