% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getvennlist.R
\name{getvennlist.default}
\alias{getvennlist.default}
\title{generate a vennlist for VennDiagram}
\usage{
\method{getvennlist}{default}(da, sampleinfo = NULL,
  factorNames = NULL, ...)
}
\arguments{
\item{da}{dataframe; a dataframe contained one character column and the others are numeric.
all columns should be numeric if sampleinfo isn't NULL.}

\item{sampleinfo}{dataframe; a sample information, default is NULL.}

\item{factorNames}{character, a column name of sampleinfo,
when sampleinfo isn't NULL, factorNames shouldn't be NULL, default is NULL.}

\item{...;}{Additional arguments passed to \code{\link[MicrobitaProcess]{CountOrRatios}}.}
}
\description{
generate a vennlist as the input of \code{[VennDiagram]}.
}
\author{
Shuangbin Xu
}
