% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{getvennlist.phyloseq}
\alias{getvennlist.phyloseq}
\title{getvennlist}
\usage{
\method{getvennlist}{phyloseq}(obj, factorNames, ...)
}
\arguments{
\item{obj}{the phyloseq object.}

\item{...}{additional parameters.}

\item{factorNamesIndex}{integer, the index of column names of sample_data.}
}
\description{
getvennlist
}
