% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{ggbartax.phyloseq}
\alias{ggbartax.phyloseq}
\title{ggbartax}
\usage{
\method{ggbartax}{phyloseq}(obj, ...)
}
\arguments{
\item{obj}{the phyloseq object.}

\item{...}{additional parameters.}

\item{sampleda}{data.frame, nrow sample * ncol factor.}
}
\description{
ggbartax
}
