% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{ggbartax}
\alias{ggbartax}
\title{ggbartax}
\usage{
ggbartax(obj, ...)
}
\arguments{
\item{obj}{data.frame or phyloseq.}

\item{...}{additional parameters}
}
\value{
barplot of tax
}
\description{
ggbartax
}
