% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdiffAnalysis.R
\docType{methods}
\name{ggdifftaxbar}
\alias{ggdifftaxbar}
\alias{ggdifftaxbar,diffAnalysisClass-method}
\alias{ggdifftaxbar,diffAnalysisClass}
\alias{ggdifftaxbar.featureMeanMedian}
\title{significantly discriminative feature barplot}
\usage{
ggdifftaxbar(obj, ...)

\S4method{ggdifftaxbar}{diffAnalysisClass}(obj, filepath = NULL,
  output = "biomarker_barplot", figwidth = 6, figheight = 3, ...)

\method{ggdifftaxbar}{featureMeanMedian}(feMeanMedian, featurename, class,
  subclass, factorLevels = NULL, setColors = TRUE, coloslist = NULL,
  settheme = TRUE, ...)
}
\arguments{
\item{obj}{object, diffAnalysisClass see also \code{\link[MicrobiotaProcess]{diffAnalysis}} 
or featureMeanMedian see also \code{\link[MicrobiotaProcess]{getMeanMedian}}.}

\item{filepath}{character, default is NULL, meaning current path.}

\item{output}{character, the output dir name, default is "biomarker_barplot".}

\item{figwidth}{numeric, the width of figures, default is 6.}

\item{figheight}{numeric, the height of figures, default is 3.}

\item{featurename}{character, the feature name, contained at the objet.}

\item{class}{character, factor name.}

\item{subclass}{character, factor name.}

\item{setColors}{logical, whether set the colors, default is TRUE, or FALSE,then
use scale_color_manual setting.}

\item{settheme}{logical, whether set the theme, default is TRUE, or FALSE, then
use theme and guides setting.}
}
\description{
significantly discriminative feature barplot
}
\author{
Shuangbin Xu
}
