% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdiffAnalysis.R
\name{ggeffectsize}
\alias{ggeffectsize}
\alias{ggeffectsize.data.frame}
\alias{ggeffectsize.diffAnalysisClass}
\title{visualization of effect size by the Linear Discriminant Analysis or randomForest}
\usage{
ggeffectsize(obj, ...)

\method{ggeffectsize}{data.frame}(data, factorName, effectsizename,
  factorLevels = NULL, setColors = TRUE, settheme = TRUE, ...)

\method{ggeffectsize}{diffAnalysisClass}(obj, removeUnkown = TRUE, ...)
}
\arguments{
\item{obj}{object, diffAnalysisClass see \code{\link[MicrobiotaProcess]{diffAnalysis}}}

\item{data}{data.frame, data.frame contained effect size and the group information.}

\item{factorName}{character, the column name contained group information in data.frame.}

\item{effectsizename}{character, the column name contained effect size information.}

\item{setColors}{logical, whether set the colors, default is TRUE, or FALSE,then 
use scale_color_manual setting.}

\item{settheme}{logical, whether set the theme, default is TRUE, or FALSE, then
use theme and guides setting.}

\item{removeUnkown}{logical, whether do not show unkown taxonomy, default is TRUE.}
}
\description{
visualization of effect size by the Linear Discriminant Analysis or randomForest
}
\author{
Shuangbin Xu
}
