% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareplot.R
\name{ggrarecurve.default}
\alias{ggrarecurve.default}
\title{Rarefaction alpha index}
\usage{
\method{ggrarecurve}{default}(data, nrows, mapping = NULL,
  linesize = 0.5, chunks = 400, sampleda, factorNames, factorlevels,
  indexNames = "Observe", se = FALSE, method = "lm", formula = y ~
  log(x), ...)
}
\arguments{
\item{data}{data.frame,(nrow sample * ncol feature (factor)) or
the data.frame for stat_smooth.}

\item{nrows, }{the nrow of facet.}

\item{mapping, }{set of aesthetic mapping of ggplot2, default is NULL,
if the data is the data.frame for stat_smooth, the mapping should be set.}

\item{linesize}{integer, default is 0.5.}

\item{chunks}{integer, the number of subsample in a sample, 
default is 400.}

\item{sampleda, }{data.frame, (nrow sample * ncol factor)}

\item{factorNames}{character, default is missing.}

\item{factorlevels}{list, the levels of the factors, default is NULL,
if you want to order the levels of factor, you can set this.}

\item{indexNames}{vector character, default is "Observe".}

\item{se}{boolean, default is FALSE.}

\item{method}{character, default is lm.}

\item{formula}{formula, default is `y ~ log(x)`}
}
\description{
building curve of alpha index in Rarefied samples.
}
\author{
ShuangbinXu
}
