% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{ggrarecurve.phyloseq}
\alias{ggrarecurve.phyloseq}
\title{ggrarecurve}
\usage{
\method{ggrarecurve}{phyloseq}(obj, ...)
}
\arguments{
\item{obj}{the phyloseq object.}

\item{...}{additional parameters.}
}
\description{
ggrarecurve
}
