% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_qiime2.R
\name{import_qiime2}
\alias{import_qiime2}
\title{Import function to load the output of qiime2.}
\usage{
import_qiime2(
  otuqza,
  taxaqza = NULL,
  mapfilename = NULL,
  refseqqza = NULL,
  treeqza = NULL,
  build_tree = FALSE,
  parallel = FALSE,
  ...
)
}
\arguments{
\item{otuqza}{character, the file contained otu table, the ouput of qiime2.}

\item{taxaqza}{character, the file contained taxonomy, the ouput of qiime2,
default is NULL.}

\item{mapfilename}{character, the file contained sample information,
the tsv format, default is NULL.}

\item{refseqqza}{character, the file contained refrentent sequences, default is
NULL.}

\item{treeqza}{character, the file contained the tree file, default is NULL.}

\item{build_tree}{logical, whether building the tree, when the rownames of 
feature table contains the sequence, default is FALSE.}

\item{parallel}{logical, whether parsing the column of taxonomy multi-parallel, default is FALSE.}

\item{..., }{additional parameters, see alos\code{\link{build_tree}}.}
}
\value{
phyloseq-class contained the argument class.
}
\description{
The function was designed to import the output of qiime2 and convert them to phyloseq
class.
}
\examples{
otuqzafile <- system.file("extdata", "table.qza",
                          package="MicrobiotaProcess")
taxaqzafile <- system.file("extdata", "taxa.qza",
                           package="MicrobiotaProcess")
mapfile <- system.file("extdata", "metadata_qza.txt",
                       package="MicrobiotaProcess")
ps <- import_qiime2(otuqza=otuqzafile, taxaqza=taxaqzafile,
                    mapfilename=mapfile)
ps
}
\author{
Shuangbin Xu
}
