% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxplot.R
\name{mappingtaxda}
\alias{mappingtaxda}
\title{generate the mapping data}
\usage{
mappingtaxda(data, topn = 30, count = FALSE, sampleda = NULL,
  factorlevels = NULL, plotda = TRUE)
}
\arguments{
\item{data}{data.frame, (nrow sample * ncol feature (factor))}

\item{topn}{integer, the top number of abundance taxonomy(feature).}

\item{count}{boolean, default is FALSE.}

\item{sampleda}{data.frame, (nrow sample * ncol factor) the 
sample information, if the data doesn't contain the information.}

\item{factorlevels}{list, the levels of the factors, default is NULL,
if you want to order the levels of factor, you can set this.}

\item{plotda}{boolean, default is TRUE, whether build the data.frame for
`geom_bar` of `ggplot2`.}
}
\description{
build the data.frame for `geom_bar` of `ggplot2`,
if the data.frame was built, you can visulize with `ggbartax`.
}
\author{
ShuangbinXu
}
