% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveSeq.R
\name{mapply_retrieve_seq}
\alias{mapply_retrieve_seq}
\title{Retriveing Sequencing from NCBI By mapply}
\usage{
mapply_retrieve_seq(
  idlist,
  files,
  databases = "protein",
  type = "fasta",
  times = 3,
  checkids = TRUE
)
}
\arguments{
\item{idlist}{vector, the accession version.}

\item{files}{character, the file name specified by a double-quoted string.}

\item{databases}{character, the name of databases to use, default is `protein`,
if nucleotide sequences to retrieve set nuccore,see \code{\link[rentrez]{entrez_fetch}}.}

\item{type}{character, the format in which to get data,such as fasta, xml ...,
see \code{\link[rentrez]{entrez_fetch}}.}

\item{times}{integer, the time of sleeping, default is 3, 
meaning 3 seconds.}

\item{checkids}{logical, whether check the sequence of ids has been retrieved.
default is FALSE.}
}
\value{
the files of sequences downloaded by ids
}
\description{
Retriveing sequences from NCBI with the accession ids.
}
\examples{
idslist <- list(c("ADM52729.1", "AAF82637.1"), 
                c("CAA24729.1", "CAA83510.1"))
mapply_retrieve_seq(idlist=idslist,
                  files="test.fasta",
                  databases="protein",
                  type="fasta",
                  times=3,checkids=TRUE)
}
\seealso{
\code{\link[MicrobiotaProcess]{retrieve_seq}}
}
\author{
Shuangbin Xu
}
