% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGroupCompare.R
\name{mapplyMultiGroupCompare}
\alias{mapplyMultiGroupCompare}
\title{multiple comparison for mulitple measures with Kruskal Wallis test}
\usage{
mapplyMultiGroupCompare(data, measurelist = NULL, padjust = "fdr", ...)
}
\arguments{
\item{data}{data frame; n(row) sample * m(col) feature and a factor col.}

\item{measurelist}{vector character; must be the names of feature, 
default is NULL,which perform the test for the total features.}

\item{padjust}{character; Method for adjusting p values, see 
p.adjust, default is `fdr`.}

\item{...;}{Additional arguments passed to 
\code{\link[MicrobitaProcess]{MultiGroupCompare}}.}
}
\description{
Multiple comparision for multiple measures with Kruskal Wallis test. which 
based on the \code{\link[MicrobitaProcess]{MultiGroupCompare}}
}
\author{
Shuangbin Xu
}
