% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveSeq.R
\name{mapplyretrieveSeq}
\alias{mapplyretrieveSeq}
\title{Retriveing Sequencing from NCBI By mapply}
\usage{
mapplyretrieveSeq(idlist, files, times = 3, checkids = TRUE)
}
\arguments{
\item{files}{character, the file name specified by a double-quoted string.}

\item{times}{integer, the time of sleeping, default is 3, 
meaning 3 seconds.}

\item{checkids}{bool, whether check the sequence of ids has been retrieved.
default is FALSE.}

\item{ids}{list, the accession number or accession.}
}
\description{
Retriveing sequences from NCBI with the accession ids.
}
\seealso{
[retrieveSeq]
}
\author{
ShuangbinXu
}
