% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoGroupTest.R
\name{mapplyTwoGroupTest}
\alias{mapplyTwoGroupTest}
\title{multiple comparison for multiple measure with tow sample test}
\usage{
mapplyTwoGroupTest(methodnames = "kruskal.test", measurelist = NULL,
  data, ...)
}
\arguments{
\item{methodnames}{character; the method of tow sample test, default is 
`kruskal.test`, one of the `t.test`, `kruskal.test` and `wilcox.test`.}

\item{measurelist}{vector; must be the names of the features, default is NULL,
meaning the total features.}

\item{data}{dataframe; the data contaiend the features and the group for
the corresponding elements of `features`, shape(`row_sample * col_feature`).}

\item{...;}{Additional arguments passed to the method choosed.}
}
\description{
Performs two sample test to multiple measure with 
`kruskal.test`, `wilcox.test`, `t.test`.
}
\author{
Shuangbin Xu
}
