% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-binary_offspring.R
\name{mp_balance_clade}
\alias{mp_balance_clade}
\alias{mp_balance_clade,MPSE-method}
\alias{mp_balance_clade,MPSE}
\alias{mp_balance_clade,tbl_mpse-method}
\alias{mp_balance_clade,tbl_mpse}
\alias{mp_balance_clade,grouped_df_mpse-method}
\alias{mp_balance_clade,grouped_df_mpse}
\title{Calculating the balance score of internal nodes (clade) according to the mean/median abundance of their binary children tips.}
\usage{
mp_balance_clade(
  .data,
  .abundance = NULL,
  force = FALSE,
  relative = TRUE,
  balance_fun = c("geometric.mean", "mean", "median"),
  pseudonum = 0.001,
  action = "get",
  ...
)

\S4method{mp_balance_clade}{MPSE}(
  .data,
  .abundance = NULL,
  force = FALSE,
  relative = TRUE,
  balance_fun = c("geometric.mean", "mean", "median"),
  pseudonum = 0.001,
  action = "get",
  ...
)

\S4method{mp_balance_clade}{tbl_mpse}(
  .data,
  .abundance = NULL,
  force = FALSE,
  relative = TRUE,
  balance_fun = c("geometric.mean", "mean", "median"),
  pseudonum = 0.001,
  action = "get",
  ...
)

\S4method{mp_balance_clade}{grouped_df_mpse}(
  .data,
  .abundance = NULL,
  force = FALSE,
  relative = TRUE,
  balance_fun = c("geometric.mean", "mean", "median"),
  pseudonum = 0.001,
  action = "get",
  ...
)
}
\arguments{
\item{.data}{MPSE object which must contain otutree slot, required}

\item{.abundance}{the column names of abundance.}

\item{force}{logical whether calculate the (relative) abundance forcibly when the abundance
is not be rarefied, default is FALSE.}

\item{relative}{logical whether calculate the relative abundance.}

\item{balance_fun}{function the method to calculate the (relative) abundance of internal nodes
according to their children tips, default is 'geometric.mean', other options are 'mean' and 'median'.}

\item{pseudonum}{numeric add a pseudo numeric to avoid the error of division in calculation, default 
is 0.001 .}

\item{action}{character, "add" joins the new information to the otutree slot if it exists (default).
In addition, "only" return a non-redundant tibble with the just new information. "get" return 'otutree'
slot, which is a treedata object.}

\item{...}{additional parameters, meaningless now.}
}
\value{
a object according to 'action' argument.
}
\description{
Calculating the balance score of internal nodes (clade) according to the mean/median abundance of their binary children tips.
}
