% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-mp_cal_cor.R
\name{mp_cal_cor}
\alias{mp_cal_cor}
\alias{mp_cal_cor,MPSE-method}
\alias{mp_cal_cor,MPSE}
\alias{mp_cal_cor,tbl_mpse-method}
\alias{mp_cal_cor,tbl_mpse}
\alias{mp_cal_cor,grouped_df_mpse-method}
\alias{mp_cal_cor,grouped_df_mpse}
\title{Calculate the correlation between the features with specified abundance.}
\usage{
mp_cal_cor(
  .data,
  .abundance,
  method = "spearman",
  alternative = c("two.sided", "less", "greater"),
  R.square = FALSE,
  p.adjust.method = NULL,
  action = "add",
  ...
)

\S4method{mp_cal_cor}{MPSE}(
  .data,
  .abundance,
  method = "spearman",
  alternative = c("two.sided", "less", "greater"),
  R.square = FALSE,
  p.adjust.method = NULL,
  action = "add",
  ...
)

\S4method{mp_cal_cor}{tbl_mpse}(
  .data,
  .abundance,
  method = "spearman",
  alternative = c("two.sided", "less", "greater"),
  R.square = FALSE,
  p.adjust.method = NULL,
  action = "add",
  ...
)

\S4method{mp_cal_cor}{grouped_df_mpse}(
  .data,
  .abundance,
  method = "spearman",
  alternative = c("two.sided", "less", "greater"),
  R.square = FALSE,
  p.adjust.method = NULL,
  action = "add",
  ...
)
}
\arguments{
\item{.data}{MPSE or tbl_mpse object}

\item{.abundance}{the name of otu abundance to be calculated}

\item{method}{character indicating which correlation coefficient to be calculate,
must be one of "pearson", "spearman", "kendall", default is "spearman".}

\item{alternative}{character indicates the alternative hypothesis,  must be one of
"two.sided", "greater" or "less", default is "two.sided".}

\item{R.square}{logical whether calculate the R square, default is FALSE.}

\item{p.adjust.method}{character correction method, default is NULL, which will 
not correct p value}

\item{action}{character, "add" joins the correlation data to the object, "only" return
a non-redundant tibble with the correlation information, "get" return 'tbl_graph' object,
which can be visualized by 'ggraph'.}

\item{...}{additional parameters.}
}
\value{
tbl_df (action='only') or tbl_graph (action='get') or MPSE object (action='add')
}
\description{
Calculate the correlation between the features with specified abundance.
}
