% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-get-pca.R
\name{mp_cal_dca}
\alias{mp_cal_dca}
\alias{mp_cal_dca,MPSE-method}
\alias{mp_cal_dca,MPSE}
\alias{mp_cal_dca,tbl_mpse-method}
\alias{mp_cal_dca,tbl_mpse}
\alias{mp_cal_dca,grouped_df_mpse-method}
\alias{mp_cal_dca,grouped_df_mpse}
\title{Detrended Correspondence Analysis with MPSE or tbl_mpse object}
\usage{
mp_cal_dca(.data, .abundance, .dim = 3, action = "add", origin = TRUE, ...)

\S4method{mp_cal_dca}{MPSE}(.data, .abundance, .dim = 3, action = "add", origin = TRUE, ...)

\S4method{mp_cal_dca}{tbl_mpse}(.data, .abundance, .dim = 3, action = "add", origin = TRUE, ...)

\S4method{mp_cal_dca}{grouped_df_mpse}(.data, .abundance, .dim = 3, action = "add", origin = TRUE, ...)
}
\arguments{
\item{.data}{MPSE or tbl_mpse object}

\item{.abundance}{the name of abundance to be calculated.}

\item{.dim}{integer The number of dimensions to be returned, default is 3.}

\item{action}{character "add" joins the 'decorana' result to the object, "only" return
a non-redundant tibble with the 'decorana' result. "get" return 'decorana' object can
be processed with related vegan function.}

\item{origin}{logical Use true origin even in detrended correspondence analysis.
default is TRUE.}

\item{...}{additional parameters see also 'vegan::decorana'}
}
\value{
update object or tbl according to the action.
}
\description{
Detrended Correspondence Analysis with MPSE or tbl_mpse object
}
