% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-get-dist.R
\name{mp_cal_dist}
\alias{mp_cal_dist}
\alias{mp_cal_dist,MPSE-method}
\alias{mp_cal_dist,MPSE}
\alias{mp_cal_dist,tbl_mpse-method}
\alias{mp_cal_dist,tbl_mpse}
\alias{mp_cal_dist,grouped_df_mpse-method}
\alias{mp_cal_dist,grouped_df_mpse}
\title{Calculate the distances between the samples with specified abundance.}
\usage{
mp_cal_dist(
  .data,
  .abundance,
  distmethod = "bray",
  action = "add",
  na.rm = TRUE,
  remove.dups = TRUE,
  ...
)

\S4method{mp_cal_dist}{MPSE}(
  .data,
  .abundance,
  distmethod = "bray",
  action = "add",
  na.rm = TRUE,
  remove.dups = FALSE,
  ...
)

\S4method{mp_cal_dist}{tbl_mpse}(
  .data,
  .abundance,
  distmethod = "bray",
  action = "add",
  na.rm = TRUE,
  remove.dups = TRUE,
  ...
)

\S4method{mp_cal_dist}{grouped_df_mpse}(
  .data,
  .abundance,
  distmethod = "bray",
  action = "add",
  na.rm = TRUE,
  remove.dups = TRUE,
  ...
)
}
\arguments{
\item{.data}{MPSE or tbl_mpse object}

\item{.abundance}{the name of otu abundance to be calculated}

\item{distmethod}{character the method to calculate distance.}

\item{action}{character, "add" joins the distance data to the object, "only" return
a non-redundant tibble with the distance information. "get" return 'dist' object.}

\item{na.rm}{logical whether to drop the rows with an NA when 'dist' object is 
converted to 'tbl_df', default is TRUE.}

\item{remove.dups}{logical whether to remove duplicate entries when 'dist' object is
converted to 'tbl_df', default is TRUE.}

\item{...}{additional parameters.}
}
\value{
update object or tibble according the 'action'
}
\description{
Calculate the distances between the samples with specified abundance.
}
