% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-get-pcoa.R
\name{mp_cal_nmds}
\alias{mp_cal_nmds}
\alias{mp_cal_nmds,MPSE-method}
\alias{mp_cal_nmds,MPSE}
\alias{mp_cal_nmds,tbl_mpse-method}
\alias{mp_cal_nmds,tbl_mpse}
\alias{mp_cal_nmds,grouped_df_mpse-method}
\alias{mp_cal_nmds,grouped_df_mpse}
\title{Nonmetric Multidimensional Scaling Analysis with MPSE or tbl_mpse object}
\usage{
mp_cal_nmds(
  .data,
  .abundance,
  distmethod = "bray",
  .dim = 2,
  action = "add",
  seed = 123,
  ...
)

\S4method{mp_cal_nmds}{MPSE}(
  .data,
  .abundance,
  distmethod = "bray",
  .dim = 2,
  action = "add",
  seed = 123,
  ...
)

\S4method{mp_cal_nmds}{tbl_mpse}(
  .data,
  .abundance,
  distmethod = "bray",
  .dim = 2,
  action = "add",
  seed = 123,
  ...
)

\S4method{mp_cal_nmds}{grouped_df_mpse}(
  .data,
  .abundance,
  distmethod = "bray",
  .dim = 2,
  action = "add",
  seed = 123,
  ...
)
}
\arguments{
\item{.data}{MPSE or tbl_mpse object}

\item{.abundance}{the name of abundance to be calculated.}

\item{distmethod}{character the method to calculate distance.}

\item{.dim}{integer The number of dimensions to be returned, default is 2.}

\item{action}{character "add" joins the NMDS result to the object, "only" return
a non-redundant tibble with the NMDS result. "get" return 'metaMDS' object can
be analyzed with related 'vegan' function.}

\item{seed}{a random seed to make this analysis reproducible, default is 123.}

\item{...}{additional parameters see also 'mp_cal_dist'.}
}
\value{
update object or tbl according to the action.
}
\description{
Nonmetric Multidimensional Scaling Analysis with MPSE or tbl_mpse object
}
