% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-get-NRI-NTI.R
\name{mp_cal_NRI_NTI}
\alias{mp_cal_NRI_NTI}
\alias{mp_cal_NRI_NTI,MPSE-method}
\alias{mp_cal_NRI_NTI,MPSE}
\alias{mp_cal_NRI_NTI,tbl_mpse-method}
\alias{mp_cal_NRI_NTI,tbl_mpse}
\alias{mp_cal_NRI_NTI,grouped_df_mpse-method}
\alias{mp_cal_NRI_NTI,grouped_df_mpse}
\title{Calculating NRI (Nearest Relative Index) and NTI (Nearest Taxon Index) with MPSE or tbl_mpse object}
\usage{
mp_cal_NRI_NTI(
  .data,
  .abundance,
  action = "add",
  abundance.weighted = TRUE,
  force = FALSE,
  seed = 123,
  ...
)

\S4method{mp_cal_NRI_NTI}{MPSE}(
  .data,
  .abundance,
  action = "add",
  abundance.weighted = TRUE,
  force = FALSE,
  seed = 123,
  ...
)

\S4method{mp_cal_NRI_NTI}{tbl_mpse}(
  .data,
  .abundance,
  action = "add",
  abundance.weighted = TRUE,
  force = FALSE,
  seed = 123,
  ...
)

\S4method{mp_cal_NRI_NTI}{grouped_df_mpse}(
  .data,
  .abundance,
  action = "add",
  abundance.weighted = TRUE,
  force = FALSE,
  seed = 123,
  ...
)
}
\arguments{
\item{.data}{object, MPSE or tbl_mpse object}

\item{.abundance}{The column name of OTU abundance column to be calculate.}

\item{action}{character it has three options, "add" joins the new information
to the input tbl (default), "only" return a non-redundant tibble with the just
new information, ang 'get' return a 'alphasample' object.}

\item{abundance.weighted}{logical, whether calculate mean nearest taxon distances for each species
weighted by species abundance, default is TRUE.}

\item{force}{logical whether calculate the alpha index even the '.abundance' is
not rarefied, default is FALSE.}

\item{seed}{integer a random seed to make the result reproducible, default is 123.}

\item{...}{additional arguments see also "ses.mpd" and "ses.mntd" of "picante".}
}
\value{
update object.
}
\description{
Calculating NRI (Nearest Relative Index) and NTI (Nearest Taxon Index) with MPSE or tbl_mpse object
}
\author{
Shuangbin Xu
}
