% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-get-pcoa.R
\name{mp_cal_pcoa}
\alias{mp_cal_pcoa}
\alias{mp_cal_pcoa,MPSE-method}
\alias{mp_cal_pcoa,MPSE}
\alias{mp_cal_pcoa,tbl_mpse-method}
\alias{mp_cal_pcoa,tbl_mpse}
\alias{mp_cal_pcoa,grouped_df_mpse-method}
\alias{mp_cal_pcoa,grouped_df_mpse}
\title{Principal Coordinate Analysis with MPSE or tbl_mpse object}
\usage{
mp_cal_pcoa(
  .data,
  .abundance,
  distmethod = "bray",
  .dim = 3,
  action = "add",
  ...
)

\S4method{mp_cal_pcoa}{MPSE}(
  .data,
  .abundance,
  distmethod = "bray",
  .dim = 3,
  action = "add",
  ...
)

\S4method{mp_cal_pcoa}{tbl_mpse}(
  .data,
  .abundance,
  distmethod = "bray",
  .dim = 3,
  action = "add",
  ...
)

\S4method{mp_cal_pcoa}{grouped_df_mpse}(
  .data,
  .abundance,
  distmethod = "bray",
  .dim = 3,
  action = "add",
  ...
)
}
\arguments{
\item{.data}{MPSE or tbl_mpse object}

\item{.abundance}{the name of abundance to be calculated.}

\item{distmethod}{character the method to calculate distance.}

\item{.dim}{integer The number of dimensions to be returned, default is 3.}

\item{action}{character "add" joins the pca result to the object, "only" return
a non-redundant tibble with the pca result. "get" return 'pcasample' object can
be visualized with 'ggordpoint'.}

\item{...}{additional parameters see also 'mp_cal_dist'.}
}
\value{
update object or tbl according to the action.
}
\description{
Principal Coordinate Analysis with MPSE or tbl_mpse object
}
