% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-get-rarecurve.R
\name{mp_cal_rarecurve}
\alias{mp_cal_rarecurve}
\alias{mp_cal_rarecurve,MPSE-method}
\alias{mp_cal_rarecurve,MPSE}
\alias{mp_cal_rarecurve,tbl_mpse-method}
\alias{mp_cal_rarecurve,tbl_mpse}
\alias{mp_cal_rarecurve,grouped_df_mpse-method}
\alias{mp_cal_rarecurve,grouped_df_mpse}
\title{Calculating the different alpha diversities index with different depth}
\usage{
mp_cal_rarecurve(.data, .abundance = NULL, chunks = 400, seed = 123, ...)

\S4method{mp_cal_rarecurve}{MPSE}(.data, .abundance = NULL, chunks = 400, seed = 123, ...)

\S4method{mp_cal_rarecurve}{tbl_mpse}(.data, .abundance = NULL, chunks = 400, seed = 123, ...)

\S4method{mp_cal_rarecurve}{grouped_df_mpse}(.data, .abundance = NULL, chunks = 400, seed = 123, ...)
}
\arguments{
\item{.data}{MPSE or tbl_mpse object}

\item{.abundance}{the name of otu abundance to be calculated.}

\item{chunks}{numeric the split number of each sample to 
calculate alpha diversity, default is 400. eg. A sample has total 40000 reads,
if chunks is 400, it will be split to 100 sub-samples (100, 200, 300,..., 40000),
then alpha diversity index was calculated based on the sub-samples.}

\item{seed}{a random seed to make the result reproducible, default is 123.}

\item{...}{additional parameters.}
}
\value{
update rarecurce calss
}
\description{
Calculating the different alpha diversities index with different depth
}
