% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-accessor.R
\name{mp_extract_abundance}
\alias{mp_extract_abundance}
\alias{mp_extract_abundance,MPSE-method}
\alias{mp_extract_abundance,MPSE}
\alias{mp_extract_abundance,tbl_mpse-method}
\alias{mp_extract_abundance,tbl_mpse}
\alias{mp_extract_abundance,grouped_df_mpse-method}
\alias{mp_extract_abundance,grouped_df_mpse}
\title{extract the abundance matrix from MPSE object or tbl_mpse object}
\usage{
mp_extract_abundance(x, .abundance, byRow = TRUE, ...)

\S4method{mp_extract_abundance}{MPSE}(x, .abundance, byRow = TRUE, ...)

\S4method{mp_extract_abundance}{tbl_mpse}(x, .abundance, byRow = TRUE, ...)

\S4method{mp_extract_abundance}{grouped_df_mpse}(x, .abundance, byRow = TRUE, ...)
}
\arguments{
\item{x}{MPSE or tbl_mpse object}

\item{.abundance}{the name of abundance to be extracted.}

\item{byRow}{logical if it is set TRUE, 'otu X sample' shape will return,
else 'sample X otu' will return.}

\item{...}{additional parameters.}
}
\value{
otu abundance a data.frame object
}
\description{
extract the abundance matrix from MPSE object or tbl_mpse object
}
