% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-accessor.R
\name{mp_extract_cor}
\alias{mp_extract_cor}
\alias{mp_extract_cor,MPSE-method}
\alias{mp_extract_cor,MPSE}
\alias{mp_extract_cor,tbl_mpse-method}
\alias{mp_extract_cor,tbl_mpse}
\alias{mp_extract_cor,grouped_df_mpse-method}
\alias{mp_extract_cor,grouped_df_mpse}
\title{Extract the correlation result of mp_cal_cor with action='add'}
\usage{
mp_extract_cor(x, method = "spearman", addtaxa = TRUE, ...)

\S4method{mp_extract_cor}{MPSE}(x, method = "spearman", addtaxa = TRUE, ...)

\S4method{mp_extract_cor}{tbl_mpse}(x, method = "spearman", addtaxa = TRUE, ...)

\S4method{mp_extract_cor}{grouped_df_mpse}(x, method = "spearman", addtaxa = TRUE, ...)
}
\arguments{
\item{x}{MPSE or tbl_mpse object.}

\item{method}{the method to calculate the correlation, default is 'spearman'}

\item{addtaxa}{logical whether add the taxonomy to the nodes}

\item{...}{additional parameters}
}
\value{
tbl_graph class containing correlation information
}
\description{
Extract the correlation result of mp_cal_cor with action='add'
}
