% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-accessor.R
\docType{methods}
\name{mp_extract_dist}
\alias{mp_extract_dist}
\alias{mp_extract_dist,MPSE-method}
\alias{mp_extract_dist,MPSE}
\alias{mp_extract_dist,tbl_mpse-method}
\alias{mp_extract_dist,tbl_mpse}
\alias{mp_extract_dist,grouped_df_mpse-method}
\alias{mp_extract_dist,grouped_df_mpse}
\title{extract the dist object from MPSE or tbl_mpse object}
\usage{
mp_extract_dist(x, distmethod, env.flag = FALSE, .group = NULL)

\S4method{mp_extract_dist}{MPSE}(x, distmethod, env.flag = FALSE, .group = NULL)

\S4method{mp_extract_dist}{tbl_mpse}(x, distmethod, env.flag = FALSE, .group = NULL)

\S4method{mp_extract_dist}{grouped_df_mpse}(x, distmethod, env.flag = FALSE, .group = NULL)
}
\arguments{
\item{x}{MPSE object or tbl_mpse object}

\item{distmethod}{character the method of calculated distance.}

\item{env.flag}{logical whether extract the distance of samples calculated
based on continuous environment factors, default is FALSE.}

\item{.group}{the column name of sample information, default is NULL, when it
is provided, a tibble that can be visualized via ggplot2 will return.}
}
\value{
dist object or tbl_df object when .group is provided.
}
\description{
extract the dist object from MPSE or tbl_mpse object
}
