% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-accessor.R
\docType{methods}
\name{mp_extract_feature}
\alias{mp_extract_feature}
\alias{mp_extract_feature,MPSE-method}
\alias{mp_extract_feature,MPSE}
\alias{mp_extract_feature,tbl_mpse-method}
\alias{mp_extract_feature,tbl_mpse}
\alias{mp_extract_feature,grouped_df_mpse-method}
\alias{mp_extract_feature,grouped_df_mpse}
\title{extract the feature (OTU) information in MPSE object}
\usage{
mp_extract_feature(x, addtaxa = FALSE, ...)

\S4method{mp_extract_feature}{MPSE}(x, addtaxa = FALSE, ...)

\S4method{mp_extract_feature}{tbl_mpse}(x, addtaxa = FALSE, ...)

\S4method{mp_extract_feature}{grouped_df_mpse}(x, addtaxa = FALSE, ...)
}
\arguments{
\item{x}{MPSE object}

\item{addtaxa}{logical whether adding the taxonomy information
default is FALSE.}

\item{...}{additional arguments}
}
\value{
tbl_df contained feature (OTU) information.
}
\description{
extract the feature (OTU) information in MPSE object
}
