% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-accessor.R
\name{mp_extract_internal_attr}
\alias{mp_extract_internal_attr}
\alias{mp_extract_internal_attr,MPSE-method}
\alias{mp_extract_internal_attr,MPSE}
\alias{mp_extract_internal_attr,tbl_mpse-method}
\alias{mp_extract_internal_attr,tbl_mpse}
\alias{mp_extract_internal_attr,grouped_df_mpse-method}
\alias{mp_extract_internal_attr,grouped_df_mpse}
\title{Extracting the PCA, PCoA, etc results from MPSE or tbl_mpse object}
\usage{
mp_extract_internal_attr(x, name, ...)

\S4method{mp_extract_internal_attr}{MPSE}(x, name, ...)

\S4method{mp_extract_internal_attr}{tbl_mpse}(x, name, ...)

\S4method{mp_extract_internal_attr}{grouped_df_mpse}(x, name, ...)
}
\arguments{
\item{x}{MPSE or tbl_mpse object}

\item{name}{character 'PCA' or 'PCoA'}

\item{...}{additional parameters}
}
\value{
prcomp or pcoa etc object
}
\description{
Extracting the PCA, PCoA, etc results from MPSE or tbl_mpse object
}
