% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-accessor.R
\name{mp_extract_refseq}
\alias{mp_extract_refseq}
\alias{mp_extract_refseq,MPSE-method}
\alias{mp_extract_refseq,MPSE}
\alias{mp_extract_refseq,tbl_mpse-method}
\alias{mp_extract_refseq,tbl_mpse}
\alias{mp_extract_refseq,grouped_df_mpse-method}
\alias{mp_extract_refseq,grouped_df_mpse}
\title{Extract the representative sequences from MPSE object}
\usage{
mp_extract_refseq(x, ...)

\S4method{mp_extract_refseq}{MPSE}(x, ...)

\S4method{mp_extract_refseq}{tbl_mpse}(x, ...)

\S4method{mp_extract_refseq}{grouped_df_mpse}(x, ...)
}
\arguments{
\item{x}{MPSE object}

\item{...}{additional parameters, meaningless now.}
}
\description{
Extract the representative sequences from MPSE object
}
