% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-accessor.R
\docType{methods}
\name{mp_extract_sample}
\alias{mp_extract_sample}
\alias{mp_extract_sample,MPSE-method}
\alias{mp_extract_sample,MPSE}
\alias{mp_extract_sample,tbl_mpse-method}
\alias{mp_extract_sample,tbl_mpse}
\alias{mp_extract_sample,grouped_df_mpse-method}
\alias{mp_extract_sample,grouped_df_mpse}
\title{extract the sample information in MPSE object}
\usage{
mp_extract_sample(x, ...)

\S4method{mp_extract_sample}{MPSE}(x, ...)

\S4method{mp_extract_sample}{tbl_mpse}(x, ...)

\S4method{mp_extract_sample}{grouped_df_mpse}(x, ...)
}
\arguments{
\item{x}{MPSE object}

\item{...}{additional arguments}
}
\value{
tbl_df contained sample information.
}
\description{
extract the sample information in MPSE object
}
