% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-mp_fortify.R
\name{mp_fortify}
\alias{mp_fortify}
\title{mp_fortify}
\usage{
mp_fortify(model, ...)
}
\arguments{
\item{model}{object}

\item{...}{additional parameters}
}
\value{
data frame or tbl_df object
}
\description{
Fortify a model with data in MicrobiotaProcess
}
