% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-io.R
\name{mp_import_qiime}
\alias{mp_import_qiime}
\title{Import function to load the output of qiime.}
\usage{
mp_import_qiime(
  otufilename,
  mapfilename = NULL,
  otutree = NULL,
  refseq = NULL,
  ...
)
}
\arguments{
\item{otufilename}{character, the file contained otu table, the ouput of qiime.}

\item{mapfilename}{character, the file contained sample information,
the tsv format, default is NULL.}

\item{otutree}{treedata, phylo or character, the file contained reference sequences, or
treedata object, which is the result by parsing function of treeio, default is NULL.}

\item{refseq}{XStringSet or character, the file contained the tree file or XStringSet, default is NULL.}

\item{..., }{additional parameters.}
}
\value{
MPSE-class.
}
\description{
The function was designed to import the output of qiime and convert them to MPSE
class.
}
\author{
Shuangbin Xu
}
