% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{mp_plot_abundance}
\alias{mp_plot_abundance}
\alias{mp_plot_abundance,MPSE-method}
\alias{mp_plot_abundance,MPSE}
\alias{mp_plot_abundance,tbl_mpse-method}
\alias{mp_plot_abundance,tbl_mpse}
\alias{mp_plot_abundance,grouped_df_mpse-method}
\alias{mp_plot_abundance,grouped_df_mpse}
\title{plotting the abundance of taxa via specified taxonomy class}
\usage{
mp_plot_abundance(
  .data,
  .abundance = NULL,
  .group = NULL,
  .sec.group = NULL,
  taxa.class = NULL,
  topn = 10,
  relative = TRUE,
  force = FALSE,
  plot.group = FALSE,
  ...
)

\S4method{mp_plot_abundance}{MPSE}(
  .data,
  .abundance = NULL,
  .group = NULL,
  .sec.group = NULL,
  taxa.class = NULL,
  topn = 10,
  relative = TRUE,
  force = FALSE,
  plot.group = FALSE,
  ...
)

\S4method{mp_plot_abundance}{tbl_mpse}(
  .data,
  .abundance = NULL,
  .group = NULL,
  .sec.group = NULL,
  taxa.class = NULL,
  topn = 10,
  relative = TRUE,
  force = FALSE,
  plot.group = FALSE,
  ...
)

\S4method{mp_plot_abundance}{grouped_df_mpse}(
  .data,
  .abundance = NULL,
  .group = NULL,
  .sec.group = NULL,
  taxa.class = NULL,
  topn = 10,
  relative = TRUE,
  force = FALSE,
  plot.group = FALSE,
  ...
)
}
\arguments{
\item{.data}{MPSE object or tbl_mpse object}

\item{.abundance}{the column name of abundance to be plotted.}

\item{.group}{the column name of group to be calculated and plotted,
default is NULL.}

\item{.sec.group}{the column name of second group to be plotted with nested facet, 
default is NULL.}

\item{taxa.class}{name of taxonomy class, default is NULL, meaning the
Phylum class will be plotted.}

\item{topn}{integer the number of the top most abundant, default is 10.}

\item{relative}{logical whether calculate the relative abundance and plotted.}

\item{force}{logical whether calculate the relative abundance forcibly
when the abundance is not be rarefied, default is FALSE.}

\item{plot.group}{logical whether plotting the abundance of specified taxa.class 
taxonomy with group not sample level, default is FALSE.}

\item{...}{additional parameters, meaningless now.}
}
\description{
plotting the abundance of taxa via specified taxonomy class
}
\examples{
\dontrun{
data(mouse.time.mpse)
mouse.time.mpse \%<>\%
  mp_rrarefy()
mouse.time.mpse
mouse.time.mpse \%<>\%
  mp_cal_abundance(.abundance=RareAbundance, action="add") \%>\%
  mp_cal_abundance(.abundance=RareAbundance, .group=time, action="add")
mouse.time.mpse
p1 <- mouse.time.mpse \%>\%
      mp_plot_abundance(.abundance=RelRareAbundanceBySample, 
                        .group=time, 
                        taxa.class="Phylum", 
                        topn=20)
p2 <- mouse.time.mpse \%>\%
      mp_plot_abundance(.abundance = Abundance,
                        taxa.class = Phylum,
                        topn = 20,
                        relative = FALSE,
                        force = TRUE
                       )
p3 <- mouse.time.mpse \%>\%
      mp_plot_abundance(.abundance = RareAbundance, 
                        .group = time,
                        taxa.class = Phylum, 
                        topn = 20,
                        relative = FALSE,
                        force = TRUE
                        )
p4 <- mouse.time.mpse \%>\%
      mp_plot_abundance(.abundance = RareAbundance,
                        .group = time,
                        taxa.class = Phylum,
                        topn = 20,
                        relative = FALSE,
                        force = TRUE,
                        plot.group = TRUE
                        )
}
}
\author{
Shuangbin Xu
}
