% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{mp_plot_upset}
\alias{mp_plot_upset}
\alias{mp_plot_upset,MPSE-method}
\alias{mp_plot_upset,MPSE}
\alias{mp_plot_upset,tbl_mpse-method}
\alias{mp_plot_upset,tbl_mpse}
\alias{mp_plot_upset,grouped_df_mpse-method}
\alias{mp_plot_upset,grouped_df_mpse}
\title{Plotting the different number of OTU between group via UpSet plot}
\usage{
mp_plot_upset(.data, .group, .upset = NULL, ...)

\S4method{mp_plot_upset}{MPSE}(.data, .group, .upset = NULL, ...)

\S4method{mp_plot_upset}{tbl_mpse}(.data, .group, .upset = NULL, ...)

\S4method{mp_plot_upset}{grouped_df_mpse}(.data, .group, .upset = NULL, ...)
}
\arguments{
\item{.data}{MPSE obejct or tbl_mpse object}

\item{.group}{the column name of group}

\item{.upset}{the column name of result after run \code{mp_cal_upset}}

\item{...}{additional parameters, meaningless now.}
}
\description{
Plotting the different number of OTU between group via UpSet plot
}
\examples{
\dontrun{
data(mouse.time.mpse)
mpse <- mouse.time.mpse \%>\%
        mp_rrarefy(.abundance=Abundance) \%>\%
        mp_cal_upset(.abundance=RareAbundance, .group=time) 
mpse
p <- mpse \%>\% mp_plot_upset(.group=time, .upset=ggupsetOftime)
p
}
}
\author{
Shuangbin Xu
}
