% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-accessor.R
\name{mp_select_as_tip}
\alias{mp_select_as_tip}
\alias{mp_select_as_tip,MPSE-method}
\alias{mp_select_as_tip,MPSE}
\alias{mp_select_as_tip,tbl_mpse-method}
\alias{mp_select_as_tip,tbl_mpse}
\alias{mp_select_as_tip,grouped_df_mpse-method}
\alias{mp_select_as_tip,grouped_df_mpse}
\title{select specific taxa level as rownames of MPSE}
\usage{
mp_select_as_tip(x, tip.level = "OTU")

\S4method{mp_select_as_tip}{MPSE}(x, tip.level = "OTU")

\S4method{mp_select_as_tip}{tbl_mpse}(x, tip.level = "OTU")

\S4method{mp_select_as_tip}{grouped_df_mpse}(x, tip.level = "OTU")
}
\arguments{
\item{x}{MPSE object}

\item{tip.level}{the taxonomy level, default is 'OTU'.}
}
\description{
select specific taxa level as rownames of MPSE
}
\examples{
\dontrun{
data(mouse.time.mpse)
newmpse <- mouse.time.mpse \%>\%
           mp_select_as_tip(tip.level = Species)
newmpse
}
}
