% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-accessor.R
\docType{methods}
\name{mp_taxatree}
\alias{mp_taxatree}
\alias{mp_taxatree,MPSE-method}
\alias{mp_taxatree,MPSE}
\alias{mp_taxatree,tbl_mpse-method}
\alias{mp_taxatree,tbl_mpse}
\alias{mp_taxatree,grouped_df_mpse-method}
\alias{mp_taxatree,grouped_df_mpse}
\title{extract the taxonomy tree in MPSE object}
\usage{
mp_taxatree(x, ...)

\S4method{mp_taxatree}{MPSE}(x, ...)

\S4method{mp_taxatree}{tbl_mpse}(x, ...)

\S4method{mp_taxatree}{grouped_df_mpse}(x, ...)
}
\arguments{
\item{x}{MPSE object}

\item{...}{additional arguments}
}
\value{
taxatree treedata object
}
\description{
extract the taxonomy tree in MPSE object
}
