% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{MPSE-class}
\alias{MPSE-class}
\title{MPSE class}
\description{
MPSE class
}
\section{Slots}{

\describe{
\item{\code{otutree}}{A treedata object of tidytree package or NULL.}

\item{\code{refseq}}{A XStringSet object of Biostrings package or NULL.}

\item{\code{...}}{Other slots from \code{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment}}}
}}

