% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\name{MPSE}
\alias{MPSE}
\title{Construct a MPSE object}
\usage{
MPSE(otutree = NULL, taxatree = NULL, refseq = NULL, ...)
}
\arguments{
\item{otutree}{A treedata object of tidytree package}

\item{taxatree}{A treedata object of tidytree package}

\item{refseq}{A XStingSet object of Biostrings package}

\item{...}{additional parameters, see also the usage 
of \code{\link[SummarizedExperiment]{SummarizedExperiment}}.}
}
\value{
MPSE object
}
\description{
Construct a MPSE object
}
