% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-utilities.R
\name{multi.compare}
\alias{multi.compare}
\title{a container for performing two or more sample test.}
\usage{
multi.compare(fun = wilcox.test, data, feature, factorNames, ...)
}
\arguments{
\item{fun}{character, the method for test, optional ""}

\item{data}{data.frame, nrow sample * ncol feature+factorNames.}

\item{feature}{vector, the features wanted to test.}

\item{factorNames}{character, the name of a factor giving the corresponding groups.}

\item{..., }{additional arguments for fun.}
}
\description{
a container for performing two or more sample test.
}
\author{
ShuangbinXu
}
