% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGroupCompare.R
\name{MultiGroupCompare}
\alias{MultiGroupCompare}
\title{multiple comparison with Kruskal Wallis test}
\usage{
MultiGroupCompare(data, measure, padj = "fdr", ...)
}
\arguments{
\item{data}{data frame; n(row) sample * m(col) feature and a factor col.}

\item{measure}{character; must be the names of feature.}

\item{padj}{character; Method for adjusting p values, see p.adjust, default is `fdr`.}

\item{...;}{Additional arguments passed to \code{\link[agricolae]{kruskal}}.}
}
\description{
Multiple comparision with Kruskal Wallis test. which based on the
\code{\link[agricolae]{kruskal}}
}
\author{
Shuangbin Xu
}
