% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ordplotClass-class}
\alias{ordplotClass-class}
\title{ordplotClass class}
\description{
ordplotClass class
}
\section{Slots}{

\describe{
\item{\code{coord}}{matrix object contained the coordinate 
for ordination plot.}

\item{\code{xlab}}{character object contained the text of xlab
for ordination plot.}

\item{\code{ylab}}{character object contained the text of ylab
for ordination plot.}

\item{\code{title}}{character object contained the text of title
for ordination plot.}
}}

