% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-show.R
\name{print}
\alias{print}
\alias{print.MPSE}
\alias{print.tbl_mpse}
\alias{print.grouped_df_mpse}
\alias{print.rarecurve}
\title{print some objects}
\usage{
\method{print}{MPSE}(
  x,
  ...,
  n = NULL,
  width = NULL,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)

\method{print}{tbl_mpse}(x, ..., n = NULL, width = NULL, max_extra_cols = NULL)

\method{print}{grouped_df_mpse}(x, ..., n = NULL, width = NULL, max_extra_cols = NULL)

\method{print}{rarecurve}(x, ..., n = NULL, width = NULL, max_extra_cols = NULL)
}
\arguments{
\item{x}{Object to format or print.}

\item{...}{Other arguments passed on to individual methods.}

\item{n}{Number of rows to show. If `NULL`, the default, will print all rows
if less than option `tibble.print_max`. Otherwise, will print
`tibble.print_min` rows.}

\item{width}{Width of text output to generate. This defaults to `NULL`, which
means use `getOption("tibble.width")` or (if also `NULL`)
`getOption("width")`; the latter displays only the columns that fit on one
screen. You can also set `options(tibble.width = Inf)` to override this
default and always print all columns.}

\item{max_extra_cols}{Number of extra columns to print abbreviated information for,
if the width is too small for the entire tibble. If `NULL`, the default,
will print information about at most `tibble.max_extra_cols` extra columns.}

\item{max_footer_lines}{integer maximum number of lines for the footer.}
}
\value{
print information
}
\description{
print some objects
}
