% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveSeq.R
\name{retrieveSeq}
\alias{retrieveSeq}
\title{Retriveing Sequencing from NCBI}
\usage{
retrieveSeq(ids, files, times = 3, checkids = FALSE)
}
\arguments{
\item{ids}{vector, the accession number or accession.}

\item{files}{character, the file name specified by a double-quoted string.}

\item{times}{integer, the time of sleeping, default is 3, 
meaning 3 seconds.}

\item{checkids}{bool, whether check the sequence of ids has been retrieved.
default is FALSE.}
}
\description{
Retriveing sequences from NCBI with the accession ids.
}
\author{
ShuangbinXu
}
