% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-rrarefy.R
\docType{methods}
\name{rrarefy}
\alias{rrarefy}
\alias{rrarefy,matrix-method}
\alias{rrarefy,matrix}
\alias{rrarefy,data.frame-method}
\alias{rrarefy,data.frame}
\alias{rrarefy,phyloseq-method}
\alias{rrarefy,phyloseq}
\alias{rrarefy,tbl_ps-method}
\alias{rrarefy,tbl_ps}
\title{rrarefy method}
\source{
rrarefy for matrix object is a wrapper method of vegan::rrarefy from the vegan 
package.
}
\usage{
rrarefy(obj, raresize, trimOTUs = TRUE, seed = 123)

\S4method{rrarefy}{matrix}(obj, raresize, trimOTUs = TRUE, seed = 123)

\S4method{rrarefy}{data.frame}(obj, raresize, trimOTUs = TRUE, seed = 123)

\S4method{rrarefy}{phyloseq}(obj, raresize, trimOTUs = TRUE, seed = 123)

\S4method{rrarefy}{tbl_ps}(obj, raresize, trimOTUs = TRUE, seed = 123)
}
\arguments{
\item{obj}{phyloseq or tbl_ps object}

\item{raresize}{integer Subsample size for rarefying community.}

\item{trimOTUs}{logical Whether to remove the otus that are no 
longer present in any sample after rarefaction}

\item{seed}{a random seed to make the rrarefy reproducible,
default is 123.}
}
\value{
update object
}
\description{
rrarefy method
}
\seealso{
\link[vegan]{rrarefy}
}
