% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml-method.R
\name{sampledflist}
\alias{sampledflist}
\title{Generate random data list from a original data.}
\usage{
sampledflist(dalist, bootnums = 30, ratio = 0.7,
  makerownames = FALSE)
}
\arguments{
\item{dalist}{list, a list contained multi data.frame.}

\item{ratio}{numeric, the ratios of each data.frame to keep.}

\item{makerownames}{logical, whether build row.names,default is FALSE.}

\item{seednums}{interger, random number generation, default is 1024.}
}
\description{
Generate random data list from a original data.
}
\author{
Shuangbin Xu
}
