% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-mp_diff_analysis.R
\name{scale_fill_diff_cladogram}
\alias{scale_fill_diff_cladogram}
\title{Create the scale of mp_plot_diff_cladogram.}
\usage{
scale_fill_diff_cladogram(values, breaks = waiver(), na.value = "grey50", ...)
}
\arguments{
\item{values}{a set of aesthetic values (different group (default)) to map data values to.}

\item{breaks}{One of 'NULL' for no breaks, ‘waiver()’ for the default breaks, A character vector of breaks.}

\item{na.value}{The aesthetic value to use for missing (‘NA’) values.}

\item{...}{see also 'discrete_scale' of 'ggplot2'.}
}
\description{
Create the scale of mp_plot_diff_cladogram.
}
