% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-mp_diff_analysis.R
\name{set_diff_boxplot_color}
\alias{set_diff_boxplot_color}
\title{set the color scale of plot generated by mp_plot_diff_boxplot}
\usage{
set_diff_boxplot_color(.data, values, ...)
}
\arguments{
\item{.data}{the aplot object generated by mp_plot_diff_boxplot.}

\item{values}{the color vector, required.}

\item{...}{additional parameters, see also the 'scale_fill_manual' of 'ggplot2'}
}
\description{
set the color scale of plot generated by mp_plot_diff_boxplot
}
