% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{set_scale_theme}
\alias{set_scale_theme}
\title{adjust the color of heatmap of mp_plot_dist}
\usage{
set_scale_theme(.data, x, aes_var)
}
\arguments{
\item{.data}{the plot of heatmap of mp_plot_dist}

\item{x}{the scale or theme}

\item{aes_var}{character the variable (column) name of color or size.}
}
\description{
adjust the color of heatmap of mp_plot_dist
}
