% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show}
\alias{show}
\title{show method}
\usage{
show(object)
}
\arguments{
\item{object}{`diffAnalysisClass` object}
}
\value{
print info
}
\description{
show method
}
\author{
Shuangbin Xu
}
