% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitStrtoList.R
\name{splitStrtoList}
\alias{splitStrtoList}
\title{split a dataframe contained one column}
\usage{
splitStrtoList(strdataframe, prefix = "tax", sep = "; ",
  extra = "drop", fill = "right", ...)
}
\arguments{
\item{strdataframe}{dataframe; a dataframe contained one column to split.}

\item{prefix}{character; the result dataframe columns names prefix, default is "tax".}

\item{sep}{character; the field separator character, default is "; ".}

\item{extra}{character; See \code{[tidyr]{separate}} details.}

\item{fill}{character; See \code{[tidyr]{separate}} details.}

\item{...;}{Additional arguments passed to \code{[tidyr]{separate}}.}
}
\description{
split a dataframe contained one column with a specify field separator character.
}
\author{
Shuangbin Xu
}
