% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareplot.R
\name{stat_rare}
\alias{stat_rare}
\title{mapping data of ggrarecurve}
\usage{
stat_rare(data, chunks = 400, sampleda, factorLevels, plotda = TRUE)
}
\arguments{
\item{data}{data.frame,(nrow sample * ncol taxonomy 
(feature) or and factor)}

\item{chunks}{integer, the number of subsample in a sample,
default is 400.}

\item{sampleda}{data.frame, (nrow sample * ncol factor)}

\item{factorLevels}{list, the levels of the factors, default is NULL,
if you want to order the levels of factor, you can set this.}

\item{plotda}{boolean, default is TRUE, whether build the data.frame for
`geom_bar` of `ggplot2`.}
}
\value{
data.frame for ggrarecurve.
}
\description{
generating the data of ggrarecurve.
}
\author{
ShuangbinXu
}
\keyword{internal}
