% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxplot.R
\name{taxbarguildes}
\alias{taxbarguildes}
\title{legend guides for ggbartax}
\usage{
taxbarguildes(keywidth = 0.3, keyheight = 0.3, ncol = 5, ...)
}
\arguments{
\item{keywidth}{numeric, the keywidth of `guide_legend`.}

\item{keyheight}{numeric, the keyheight of `guide_legend`.}

\item{ncol}{integer, the ncol of legend.}

\item{..., }{additional parameter.}
}
\description{
the legend guides for `ggbartax`
}
\author{
ShuangbinXu
}
