% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-methods.R
\name{theme_stamp}
\alias{theme_stamp}
\title{set the theme of ggplot object with the STAMP style.}
\usage{
theme_stamp(colour = c("white", "grey90"), axis = "y", ...)
}
\arguments{
\item{colour}{character the color of theme stamp.}

\item{axis}{character which grid of axis will be filled, default is 'y'.}

\item{...}{additional parameter, see also 'theme' of 'ggplot2'.}
}
\description{
set the theme of ggplot object with the STAMP style.
}
