% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-method.R
\name{tidydiffAnalysis}
\alias{tidydiffAnalysis}
\title{get the table of diffAnalysisClass}
\usage{
tidydiffAnalysis(obj)
}
\arguments{
\item{obj}{object, diffAnalysisClass}
}
\description{
get the table of diffAnalysisClass
}
