% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoGroupTest.R
\name{TwoGroupTest}
\alias{TwoGroupTest}
\title{Two sample Test}
\usage{
TwoGroupTest(measure, methodnames = "kruskal.test", data, ...)
}
\arguments{
\item{measure}{character; must be the names of feature.}

\item{methodnames}{character; the method of tow sample test, default is
`kruskal.test`, one of the `t.test`, `kruskal.test` and `wilcox.test`.}

\item{data}{dataframe; the data contaiend the features and the group for
the corresponding elements of `features`, shape(`row_sample * col_feature`).}

\item{...;}{Additional arguments passed to the method choosed.}
}
\description{
Performs two sample test with `kruskal.test`, `wilcox.test`, `t.test`.
}
\author{
Shuangbin Xu
}
