% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaIndex.R
\name{alphaindex}
\alias{alphaindex}
\title{alpha index}
\usage{
alphaindex(data, mindepth = NULL)
}
\arguments{
\item{data}{data.frame, (nrow sample * ncol taxonomy(feature))}

\item{mindepth}{numeric, Subsample size for rarefying community.}
}
\value{
data.frame contained alpha Index.
}
\description{
caculate the alpha index (Obseve,Chao1,Shannon,Simpson) of sample
with \code{\link[vegan]{diversity}}
}
\examples{
library(tidyverse)
library(vegan)
otudafile <- system.file("extdata", "otu_tax_table.txt", 
                        package="MicrobiotaProcess")
otuda <- read.table(otudafile, sep="\\t", 
             header=TRUE, row.names=1, 
             check.names=FALSE, skip=1, comment.char="")
otuda <- otuda[sapply(otuda, is.numeric)] \%>\% t() \%>\% 
          data.frame(check.names=FALSE)
set.seed(1024)
alphatab <- alphaindex(otuda)
head(alphatab)
}
\author{
ShuangbinXu
}
