% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaIndex.R
\name{alphaindex}
\alias{alphaindex}
\title{alpha index}
\usage{
alphaindex(data, mindepth = NULL)
}
\arguments{
\item{data}{data.frame, (nrow sample * ncol taxonomy(feature))}

\item{mindepth}{numeric, Subsample size for rarefying community.}
}
\description{
caculate the alpha index (Obseve,Chao1,Shannon,Simpson) of sample
with \code{\link[vegan]{diversity}}
}
\author{
ShuangbinXu
}
