% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-as-MPSE.R
\name{as.MPSE}
\alias{as.MPSE}
\title{as.MPSE method}
\usage{
as.MPSE(.data, ...)
}
\arguments{
\item{.data}{tbl_mpse or phyloseq or TreeSummarizedExperiment object}

\item{...}{additional parameters, meaningless now.}
}
\value{
MPSE object
}
\description{
convert the .data object to MPSE object
}
\examples{
data(test_otu_data)
test_otu_data \%>\% as.MPSE -> mpse
mpse
}
\author{
Shuangbin Xu
}
