% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-as-phyloseq.R
\name{as.phyloseq}
\alias{as.phyloseq}
\alias{as_phyloseq}
\alias{as.phyloseq.tbl_mpse}
\title{convert to phyloseq object.}
\usage{
as.phyloseq(x, ...)

as_phyloseq(x, ...)

\method{as.phyloseq}{tbl_mpse}(x, ...)
}
\arguments{
\item{x}{object, tbl_mpse object, which the result of 
as_tibble for phyloseq objcet.}

\item{...}{additional params}
}
\value{
phyloseq object.
}
\description{
convert to phyloseq object.
}
