% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-as-phyloseq.R
\name{as.phyloseq}
\alias{as.phyloseq}
\alias{as_phyloseq}
\alias{as.phyloseq.MPSE}
\alias{as.phyloseq.tbl_mpse}
\title{convert to phyloseq object.}
\usage{
as.phyloseq(x, .abundance, ...)

as_phyloseq(x, .abundance, ...)

\method{as.phyloseq}{MPSE}(x, .abundance, ...)

\method{as.phyloseq}{tbl_mpse}(x, .abundance, ...)
}
\arguments{
\item{x}{object, tbl_mpse object, which the result of 
as_tibble for phyloseq objcet.}

\item{.abundance}{the column name to be as the abundance 
of otu table, default is Abundance.}

\item{...}{additional params}
}
\value{
phyloseq object.
}
\description{
convert to phyloseq object.
}
