% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.treedata.data.frame}
\alias{as.treedata.data.frame}
\title{convert dataframe contained hierarchical relationship or other classes to treedata class}
\usage{
\method{as.treedata}{data.frame}(data, ...)
}
\description{
convert dataframe contained hierarchical relationship or other classes to treedata class
}
\author{
Shuangbin Xu
}
