% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-as.treedata.R
\name{as.treedata.taxonomyTable}
\alias{as.treedata.taxonomyTable}
\title{as.treedata}
\usage{
\method{as.treedata}{taxonomyTable}(tree, ...)
}
\arguments{
\item{tree}{object, This is for taxonomyTable class, 
so it should be a taxonomyTable.}

\item{...}{additional parameters.}
}
\description{
convert taxonomyTable to treedata
}
\examples{
data(test_otu_data)
tree <- as.treedata(phyloseq::tax_table(test_otu_data))
}
