% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert_to_treedata}
\alias{convert_to_treedata}
\title{convert dataframe contained hierarchical relationship or other classes to treedata class}
\usage{
convert_to_treedata(data, ...)
}
\arguments{
\item{data}{data.frame, such like the tax_table of phyloseq.}
}
\description{
convert dataframe contained hierarchical relationship or other classes to treedata class
}
\author{
Shuangbin Xu
}
