% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountOrRatios.R
\name{CountOrRatios}
\alias{CountOrRatios}
\title{caculate the count or relative abundance of replicate element with a speficify column}
\usage{
CountOrRatios(
  data,
  featurelist,
  countmode = TRUE,
  multiplenum = 1,
  rownamekeep = FALSE
)
}
\arguments{
\item{data}{dataframe; a dataframe contained one character column and others is numeric,
if featurelist is NULL. Or a numeirc dataframe, if featurelist is non't NULL, all columns 
should be numeric.}

\item{featurelist}{dataframe; a dataframe contained one chatacter column, default is NULL.}

\item{countmode}{boolean; whether return the count results (TRUE),or relative abundance (FALSE).}

\item{multiplenum}{numeric; the multiple you want to increase,default is 1.}

\item{rownamekeep}{boolean; whether you return a dataframe contained the rownames,default is FALSE.}
}
\value{
mean of data.frame by featurelist
}
\description{
Caculate the count or relative abundance of replicate element with a speficify columns
}
\examples{
otudafile <- system.file("extdata", "otu_tax_table.txt", 
                      package="MicrobiotaProcess")
samplefile <- system.file("extdata", 
                 "sample_info.txt", package="MicrobiotaProcess")
otuda <- read.table(otudafile, sep="\t", header=TRUE, 
                    row.names=1, check.names=FALSE, 
                    skip=1, comment.char="")
sampleda <- read.table(samplefile, 
            sep="\t", header=TRUE, row.names=1)
taxdf <- otuda[!sapply(otuda, is.numeric)]
taxdf <- splitStrtoList(taxdf)
otuda <- otuda[sapply(otuda, is.numeric)]
phycount <- CountOrRatios(otuda, taxdf[,2,drop=FALSE])
phyratios <- CountOrRatios(otuda, taxdf[,2,drop=FALSE], 
                           countmode=FALSE)
}
\author{
Shuangbin Xu
}
