% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-method.R
\name{diffAnalysis}
\alias{diffAnalysis}
\alias{diffAnalysis.data.frame}
\title{Differential expression analysis based on kruskal.test, generalized fold change and wilcox.test}
\usage{
diffAnalysis(x, ...)

\method{diffAnalysis}{data.frame}(data, sampleda, class, subclass = NULL,
  padjust = "fdr", ...)
}
\description{
Differential expression analysis based on kruskal.test, generalized fold change and wilcox.test
}
\author{
Shuangbin Xu
}
