% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/droptax.R
\name{droptax}
\alias{droptax}
\title{Dropping Species with Few abundance and Few Occurrences}
\usage{
droptax(taxtab, rmode = FALSE, minocc = 0, minabu = 0)
}
\arguments{
\item{taxtab}{dataframe; a dataframe of species (or features), default is (n_sample, n_feature).}

\item{rmode}{boolean; whether transpose the taxtab, default is False.}

\item{minocc}{numeric; the threshold number of occurrences to be dropped, if < 1.0,
it will be the threshold ratios of occurrences, default is 0.}

\item{minabu}{numeric: the threshold abundance, if fewer than the threshold will be dropped, default is 0.}
}
\value{
a list contained feature dataframe dropped, and the call, arguments.
}
\description{
Drop species or features from the feature data frame that occur fewer than or equal to a threshold number of 
occurrences and fewer abundance than to a threshold abundance.
}
\examples{

library("microproccess")
data <- read.csv(system.file("data", package="microproccess", "Baxter_16_crc_genera_group.csv.gz"))
data$Group <- NULL
dim(data)
head(data)
newdat <- droptax(data, rmode=FALSE, minocc=0.2, minabu=0)
dim(newdat)
head(newdat)
}
\author{
Shuangbin Xu
}
