% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-get-NRI-NTI.R
\name{get_NRI_NTI}
\alias{get_NRI_NTI}
\alias{get_NRI_NTI,matrix-method}
\alias{get_NRI_NTI,matrix}
\alias{get_NRI_NTI,data.frame-method}
\alias{get_NRI_NTI,data.frame}
\alias{get_NRI_NTI,phyloseq-method}
\alias{get_NRI_NTI,phyloseq}
\title{NRI (Nearest Relative Index) and NTI (Nearest Taxon Index)}
\usage{
get_NRI_NTI(obj, ...)

\S4method{get_NRI_NTI}{matrix}(
  obj,
  mindepth,
  sampleda,
  tree,
  abundance.weighted = TRUE,
  force = FALSE,
  ...
)

\S4method{get_NRI_NTI}{data.frame}(obj, mindepth, sampleda, tree, abundance.weighted = TRUE, ...)

\S4method{get_NRI_NTI}{phyloseq}(obj, mindepth, abundance.weighted = TRUE, ...)
}
\arguments{
\item{obj}{object, data.frame of (nrow sample * ncol taxonomy(feature)) 
or phyloseq.}

\item{...}{additional arguments see also "ses.mpd" and "ses.mntd" of "picante".}

\item{mindepth}{numeric, Subsample size for rarefying community.}

\item{sampleda}{data.frame, sample information, row sample * column factors.}

\item{tree}{tree object, it can be phylo object or treedata object.}

\item{abundance.weighted}{logical, whether calculate mean nearest taxon distances for each species 
weighted by species abundance, default is TRUE.}

\item{force}{logical whether calculate the index even the count of otu is
not rarefied, default is FALSE. If it is TRUE, meaning the rarefaction is not be
performed automatically.}
}
\value{
alphasample object contained NRT and NTI.
}
\description{
calculate NRT and NTI of sample. It is a wrapper method of picante::ses.mpd
and picante::ses.mntd
}
\author{
Shuangbin Xu
}
