% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpca.R
\name{getcoord}
\alias{getcoord}
\alias{getcoord.prcomp}
\title{get ordination coordinates.}
\usage{
getcoord(obj)

\method{getcoord}{prcomp}(obj)
}
\arguments{
\item{obj}{prcomp object}
}
\description{
get ordination coordinates.

get ordination coordinates
}
