% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getordination.R, R/getpca.R
\name{getcoord.pcoa}
\alias{getcoord.pcoa}
\alias{getcoord}
\alias{getcoord.prcomp}
\title{get ordination coordinates.}
\usage{
\method{getcoord}{pcoa}(obj, pc)

getcoord(obj, pc)

\method{getcoord}{prcomp}(obj, pc)
}
\arguments{
\item{obj}{prcomp object}

\item{pc}{integer vector, the component index.}
}
\description{
get ordination coordinates.

get ordination coordinates
}
