% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getordination.R
\name{getdist}
\alias{getdist}
\alias{getdist.default}
\alias{getdist.phyloseq}
\title{calculate distance}
\usage{
getdist(obj, ...)

\method{getdist}{default}(data, distmethod = "euclidean",
  taxa_are_rows = FALSE, sampleda = NULL, tree = NULL,
  method = "hellinger", ...)

\method{getdist}{phyloseq}(obj, distmethod = "euclidean",
  type = "sample", method = "hellinger", ...)
}
\arguments{
\item{obj}{phyloseq, phyloseq class.}

\item{..., }{additional parameters.}

\item{data}{data.frame, nrow sample * ncol feature.}

\item{distmethod}{character, default is "euclidean", see also \code{\link[phyloseq]{distanceMethodList}}}

\item{taxa_are_rows}{logical, default is FALSE.}

\item{sampleda}{data.frame, nrow sample * ncol factors.}

\item{tree}{phylo, see also \code{\link[ape]{phylo}}.}

\item{method}{character, default is hellinger, see alse \code{\link[vegan]{decostand}}}
}
\value{
distance class contianed distmethod and originalD attr
}
\description{
calculate distance
}
\examples{
data(test_otu_data)
distclass <- getdist(test_otu_data)
hcsample <- getclust(distclass)
}
\seealso{
\code{\link[phyloseq]{distance}}
}
