% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpca.R
\name{getpca}
\alias{getpca}
\alias{getpca.default}
\alias{getpca.phyloseq}
\title{Performs a principal components analysis}
\usage{
getpca(obj, ...)

\method{getpca}{default}(data, sampleda = NULL, method = "hellinger",
  ...)

\method{getpca}{phyloseq}(obj, method = "hellinger", ...)
}
\arguments{
\item{obj}{phyloseq, phyloseq class.}

\item{...}{additional parameters, see\code{\link[stat]{prcomp}}.}

\item{data}{data.frame, shape of data.frame is nrow sample * ncol feature.}

\item{sampleda}{data.frame, nrow sample * ncol factors.}

\item{method}{character, the standardization methods for
community ecologists. see \code{\link[vegan]{decostand}}.}
}
\value{
pcasample class, contained prcomp class and sample information.
}
\description{
Performs a principal components analysis
}
\examples{
library(phyloseq)
data(GlobalPatterns)
subGlobal <- subset_samples(GlobalPatterns, 
         SampleType \%in\% c("Feces", "Mock", "Ocean", "Skin"))
pcares <- getpca(subGlobal, method="hellinger")
#pcaplot <- ggordpoint(pcares, biplot=TRUE, 
#                      speciesannot=TRUE,
#                      factorNames=c("SampleType"), ellipse=TRUE)
}
