% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getordination.R
\name{getpcoa}
\alias{getpcoa}
\alias{getpcoa.default}
\alias{getpcoa.dist}
\alias{getpcoa.phyloseq}
\title{performs principal coordinate analysis (PCoA)}
\usage{
getpcoa(obj, ...)

\method{getpca}{default}(data, distmethod = "euclidean",
  taxa_are_rows = FALSE, sampleda = NULL, tree = NULL,
  type = "sample", method = "hellinger", ...)

\method{getpcoa}{dist}(obj, distmethod, data = NULL, sampleda = NULL,
  method = "hellinger", ...)

\method{getpcoa}{phyloseq}(obj, distmethod = "euclidean", ...)
}
\description{
performs principal coordinate analysis (PCoA)
}
