% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getordination.R, R/getpca.R
\name{getvarct.pcoa}
\alias{getvarct.pcoa}
\alias{getvarct}
\alias{getvarct.prcomp}
\alias{getvarct.pcasample}
\title{get the contribution of variables}
\usage{
\method{getvarct}{pcoa}(obj, ...)

getvarct(obj, ...)

\method{getvarct}{prcomp}(obj, ...)

\method{getvarct}{pcasample}(obj, ...)
}
\arguments{
\item{obj}{prcomp class or pcasample class}

\item{...}{additional parameters.}
}
\value{
the VarContrib class, contained the 
contribution and coordinate of features.
}
\description{
get the contribution of variables
}
\examples{
library(phyloseq)
data(GlobalPatterns)
subGlobal <- subset_samples(GlobalPatterns,
         SampleType \%in\% c("Feces", "Mock", "Ocean", "Skin"))
pcares <- getpca(subGlobal, method="hellinger") 
varres <- getvarct(pcares)
}
