% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpca.R
\name{getvarct}
\alias{getvarct}
\alias{getvarct.prcomp}
\alias{getvarct.pcasample}
\title{get the contribution of variables}
\usage{
getvarct(obj, ...)

\method{getvarct}{prcomp}(obj)

\method{getvarct}{pcasample}(obj)
}
\arguments{
\item{obj}{prcomp class or pcasample class}
}
\description{
get the contribution of variables

get the contribution of variables

get the contribution of variables
}
