% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getvennlist.R
\name{getvennlist}
\alias{getvennlist}
\title{generate a vennlist for VennDiagram}
\usage{
getvennlist(da, sampleinfo = NULL, ...)
}
\arguments{
\item{da}{dataframe; a dataframe contained one character column and others is numeric 
if sampleinfo is NULL, else a numeirc dataframe.}

\item{sampleinfo}{dataframe; a sample information.}

\item{...;}{Additional arguments passed to \code{\link[MicrobitaProcess]{CountOrRatios}}.}
}
\description{
generate a vennlist as the input of \code{[VennDiagram]}.
}
\author{
Shuangbin Xu
}
