% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/getvennlist.R
\name{getvennlist}
\alias{getvennlist}
\alias{getvennlist.phyloseq}
\alias{getvennlist.default}
\title{generate a vennlist for VennDiagram}
\usage{
getvennlist(obj, ...)

\method{getvennlist}{phyloseq}(obj, factorNames, ...)

\method{getvennlist}{default}(da, sampleinfo = NULL,
  factorNames = NULL, ...)
}
\arguments{
\item{obj}{data.frame or phyloseq class, a dataframe contained one character column and the others are numeric.
all columns should be numeric if sampleinfo isn't NULL.}

\item{...}{additional parameters, see \code{\link[MicrobitaProcess]{CountOrRatios}}.}

\item{factorNames}{character, a column name of sampleinfo, 
when sampleinfo isn't NULL, factorNames shouldn't be NULL, default is NULL.}

\item{sampleinfo}{dataframe; a sample information, default is NULL.}
}
\value{
return a list for VennDiagram.
}
\description{
generate a vennlist for VennDiagram

generate a vennlist for VennDiagram
}
\author{
ShuangbinXu

Shuangbin Xu
}
