% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/taxplot.R
\name{ggbartax}
\alias{ggbartax}
\alias{ggbartax.phyloseq}
\alias{ggbartax.default}
\title{taxonomy barplot}
\usage{
ggbartax(obj, ...)

\method{ggbartax}{phyloseq}(obj, ...)

\method{ggbartax}{default}(obj, mapping = NULL, position = "stack",
  stat = "identity", width = 0.7, topn = 30, count = FALSE,
  sampleda = NULL, factorLevels = NULL, facetNames = NULL,
  plotgroup = FALSE, groupfun = mean, ...)
}
\arguments{
\item{obj}{phyloseq, phyloseq class or data.frame, 
(nrow sample * ncol feature (factor)) or the data.frame for geom_bar.}

\item{...}{additional parameters, see \code{\link[ggplot2]{ggplot}}}

\item{mapping}{set of aesthetic mapping of ggplot2, default is NULL,
if the data is the data.frame for geom_bar, the mapping should be set.}

\item{position}{character, default is `stack`.}

\item{stat}{character, default is `identity`.}

\item{width}{numeric, the width of bar, default is 0.7.}

\item{topn}{integer, the top number of abundance taxonomy(feature).}

\item{count}{logical, whether show the relative abundance.}

\item{sampleda}{data.frame, (nrow sample * ncol factor), the sample 
information, if the data doesn't contain the information.}

\item{factorLevels}{list, the levels of the factors, default is NULL,
if you want to order the levels of factor, you can set this.}

\item{facetNames}{character, default is NULL.}

\item{plotgroup}{logical, whether calculate the mean or median etc 
for each group, default is FALSE.}

\item{groupfun}{character, how to calculate for feature in each group,
the default is `mean`, this will plot the mean of feature in each group.}
}
\value{
barplot of tax
}
\description{
taxonomy barplot
}
\examples{
library(ggplot2)
data(test_otu_data)
otubar <- ggbartax(test_otu_data) + 
         xlab(NULL) + ylab("relative abundance(\%)")
}
\author{
ShuangbinXu
}
