% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/taxplot.R
\name{ggbartax}
\alias{ggbartax}
\alias{ggbartax.phyloseq}
\alias{ggbartax.default}
\title{taxonomy barplot}
\usage{
ggbartax(obj, ...)

\method{ggbartax}{phyloseq}(obj, ...)

\method{ggbartax}{default}(data, mapping = NULL, position = "stack",
  stat = "identity", width = 0.7, topn = 30, count = FALSE,
  sampleda = NULL, factorLevels = NULL, settheme = TRUE,
  facetNames = NULL, setColors = TRUE, ...)
}
\arguments{
\item{obj}{data.frame or phyloseq class. data.frame should be 
(nrow sample * ncol feature (factor)) or the data.frame for geom_bar.}

\item{...}{additional parameters, see \code{\link[ggplot2]{ggplot}}}

\item{mapping}{set of aesthetic mapping of ggplot2, default is NULL,
if the data is the data.frame for geom_bar, the mapping should be set.}

\item{position, }{default is `stack`.}

\item{stat, }{default is `identity`.}

\item{width, }{the width of bar, default is 0.7.}

\item{topn, }{the top number of abundance taxonomy(feature).}

\item{count, }{whether show the relative abundance.}

\item{sampleda}{data.frame, (nrow sample * ncol factor), the sample 
information, if the data doesn't contain the information.}

\item{factorLevels}{list, the levels of the factors, default is NULL,
if you want to order the levels of factor, you can set this.}

\item{settheme}{logical, default is TRUE, or you can set FALSE, then
set the theme by youself.}

\item{facetNames}{character, default is NULL.}

\item{setColors}{logical, default is TRUE, or you can set FALSE, then 
set colors by `scale_fill_manual` of `ggplot2`.}
}
\value{
barplot of tax
}
\description{
taxonomy barplot
}
\author{
ShuangbinXu
}
