% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdiffAnalysis.R
\name{ggdiffclade}
\alias{ggdiffclade}
\alias{ggdiffclade.data.frame}
\alias{ggdiffclade.diffAnalysisClass}
\title{plot the clade tree results of different analysis}
\usage{
ggdiffclade(obj, ...)

\method{ggdiffclade}{data.frame}(taxda, nodedf, factorName,
  layout = "circular", size = 0.6, alpha = 0.4, taxlevel = 6,
  cladetext = 2, factorLevels = NULL, setColors = TRUE,
  settheme = TRUE, setlegend = TRUE, ...)

\method{ggdiffclade}{diffAnalysisClass}(obj, ...)
}
\arguments{
\item{obj}{object, diffAnalysisClass, the results of diffAnalysis \code{\link[MicrobiotaProcess]{diffAnalysis}}.}

\item{taxda}{data.frame, data.frame contained hierarchical relationship or other classes,
such like the tax_data of phyloseq.}

\item{nodedf}{data.frame, contained the tax and the factor information and(or pvalue).}

\item{factorName}{character, the names of factor in nodedf.}

\item{layout}{character, the layout of ggtree, but only "rectangular" and "circular" in here, 
default is circular.}

\item{size}{numeric, the size of segment of ggtree, default is 0.6.}

\item{alpha}{numeric, the alpha of clade, default is 0.4.}

\item{taxlevel}{positive integer, the full text of clade, default is 5.}

\item{cladetext}{numeric, the size of text of clade, default is 2.}

\item{factorLevels}{list, the levels of the factors, default is NULL,
if you want to order the levels of factor, you can set this.}

\item{setColors}{logical, whether set the color of clade, default is TRUE, 
or set FALSE,then use 'scale_fill_manual' setting.}

\item{settheme}{logical, whether set the theme of tree, default is TRUE,
or set FALSE, then use 'theme' setting.}

\item{setlegend}{logical, whether set the legend, default is TRUE,
or set FALSE, then use 'guides' or 'theme' setting.}
}
\description{
plot the clade tree results of different analysis
}
\author{
Shuangbin Xu
}
