% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdiffAnalysis.R
\name{ggdiffclade}
\alias{ggdiffclade}
\alias{ggdiffclade.data.frame}
\alias{ggdiffclade.diffAnalysisClass}
\title{plot the clade tree with highlight}
\usage{
ggdiffclade(obj, ...)

\method{ggdiffclade}{data.frame}(obj, nodedf, factorName,
  layout = "circular", size = 0.6, skpointsize = 0.8, alpha = 0.4,
  taxlevel = 6, cladetext = 2, factorLevels = NULL,
  setColors = TRUE, settheme = TRUE, setlegend = TRUE, ...)

\method{ggdiffclade}{diffAnalysisClass}(obj, removeUnkown = TRUE, ...)
}
\arguments{
\item{obj}{object, diffAnalysisClass, the results of diffAnalysis 
see also \code{\link[MicrobiotaProcess]{diffAnalysis}}, or data.frame,
contained hierarchical relationship or other classes.}

\item{..., }{additional parameters.}

\item{nodedf}{data.frame, contained the tax and the factor 
information and(or pvalue).}

\item{factorName}{character, the names of factor in nodedf.}

\item{layout}{character, the layout of ggtree, but only "rectangular" 
and "circular" in here, default is circular.}

\item{size}{numeric, the size of segment of ggtree, default is 0.6.}

\item{skpointsize}{numeric, the point size of skeleton of tree, 
default is 0.8 .}

\item{alpha}{numeric, the alpha of clade, default is 0.4.}

\item{taxlevel}{positive integer, the full text of clade, default is 5.}

\item{cladetext}{numeric, the size of text of clade, default is 2.}

\item{factorLevels}{list, the levels of the factors, default is NULL,
if you want to order the levels of factor, you can set this.}

\item{setColors}{logical, whether set the color of clade, default is TRUE, 
or set FALSE,then use 'scale_fill_manual' setting.}

\item{settheme}{logical, whether set the theme of tree, default is TRUE,
or set FALSE, then use 'theme' setting.}

\item{setlegend}{logical, whether set the legend, default is TRUE,
or set FALSE, then use 'guides' or 'theme' setting.}

\item{removeUnkown}{logical, whether do not show unkown taxonomy, 
default is TRUE.}
}
\value{
figures of tax clade show the significant different feature.
}
\description{
plot results of different analysis or data.frame, 
contained hierarchical relationship or other classes,such like the 
tax_data of phyloseq.
}
\examples{
data(kostic2012crc)
kostic2012crc
head(phyloseq::sample_data(kostic2012crc),3)
kostic2012crc <- phyloseq::rarefy_even_depth(kostic2012crc,
                         rngseed=1024)
table(phyloseq::sample_data(kostic2012crc)$DIAGNOSIS)
set.seed(1024)
diffres <- diffAnalysis(kostic2012crc, class="DIAGNOSIS",
                        mlfun="lda", filtermod="fdr",
                        firstcomfun = "kruskal.test",
                        firstalpha=0.05, strictmod=TRUE,
                        secondcomfun = "wilcox.test",
                        submin=3, subclwilc=TRUE,
                        secondalpha=0.01, lda=3)
library(ggplot2)
diffcladeplot <- ggdiffclade(diffres,alpha=0.3, size=0.2, 
                        skpointsize=0.4, 
                        taxlevel=3,
                        settheme=TRUE,
                        setColors=FALSE) +
        scale_fill_manual(values=c('#00AED7', 
                                   '#FD9347', 
                                   '#C1E168'))
}
\author{
Shuangbin Xu
}
