% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpca.R
\name{ggordpoint}
\alias{ggordpoint}
\alias{ggordpoint.default}
\alias{ggordpoint.pcasample}
\title{ordination plotter based on ggplot2.}
\usage{
ggordpoint(obj, ...)

\method{ggordpoint}{default}(obj, pc = c(1, 2), mapping = NULL,
  sampleda = NULL, factorNames = NULL, factorLevels = NULL,
  poinsize = 2, linesize = 0.3, arrowsize = 1.5,
  arrowlinecolour = "grey", biplot = TRUE, topn = 5,
  settheme = TRUE, speciesannot = FALSE, textsize = 2.5,
  fontface = "bold.italic", fontfamily = "sans", textlinesize = 0.02,
  ...)

\method{ggordpoint}{pcasample}(obj, ...)
}
\arguments{
\item{obj}{prcomp class or pcasample class,}

\item{...}{additional parameters, see \code{\link[ggrepel]{geom_text_repel}}.}

\item{pc}{integer vector, the component index.}

\item{mapping}{set of aesthetic mapping of ggplot2, default is NULL.}

\item{sampleda}{data.frame, nrow sample * ncol factors, default is NULL.}

\item{factorNames}{vector, the names of factors contained sampleda.}

\item{factorLevels}{list, the levels of the factors, default is NULL,
if you want to order the levels of factor, you can set this.}

\item{poinsize}{numeric, the size of point, default is 2.}

\item{linesize}{numeric, the line size of segment, default is 0.3.}

\item{arrowsize}{numeric, the size of arrow, default is 1.5.}

\item{arrowlinecolour}{character, the color of segment, default is grey.}

\item{biplot}{logical, whether plot the species, default is TRUE.}

\item{topn}{integer, the number species have top important contribution,
default is 5.}

\item{settheme}{logical, whether set the theme for the plot, default is TRUE.}

\item{speciesannot}{logical, whether plot the species, default is FALSE.}

\item{textsize}{numeric, the size of text, default is 2.5.}

\item{fontface}{character, the font face, default is "blod.italic".}

\item{fontfamily}{character, the font family, default is "sans".}

\item{textlinesize}{numeric, the segment size in \code{\link[ggrepel]{geom_text_repel}}.}
}
\description{
ordination plotter based on ggplot2.
}
\author{
ShuangbinXu
}
