% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/rareplot.R
\name{ggrarecurve}
\alias{ggrarecurve}
\alias{ggrarecurve.phyloseq}
\alias{ggrarecurve.default}
\title{Rarefaction alpha index}
\usage{
ggrarecurve(obj, ...)

\method{ggrarecurve}{phyloseq}(obj, ...)

\method{ggrarecurve}{default}(
  obj,
  sampleda,
  indexNames = "Observe",
  linesize = 0.5,
  facetnrow = 1,
  mapping = NULL,
  chunks = 400,
  factorNames,
  factorLevels,
  se = FALSE,
  method = "lm",
  formula = y ~ log(x),
  ...
)
}
\arguments{
\item{obj}{phyloseq, phyloseq class or data.frame
shape of data.frame (nrow sample * ncol feature (factor)) 
or ' the data.frame for stat_smooth.}

\item{...}{additional parameters, 
see also \code{\link{ggplot2}{ggplot}}.}

\item{sampleda}{data.frame, (nrow sample * ncol factor)}

\item{indexNames}{character, default is "Observe",
only for "Observe", "Chao1", "ACE", "Shannon", "Simpson", "J".}

\item{linesize}{integer, default is 0.5.}

\item{facetnrow}{integer, the nrow of facet, default is 1.}

\item{mapping}{set of aesthetic mapping of ggplot2, default is NULL,
if the data is the data.frame for stat_smooth, the mapping should be set.}

\item{chunks}{integer, the number of subsample in a sample,
default is 400.}

\item{factorNames}{character, default is missing.}

\item{factorLevels}{list, the levels of the factors, default is NULL,
if you want to order the levels of factor, you can set this.}

\item{se}{logical, default is FALSE.}

\item{method}{character, default is lm.}

\item{formula}{formula, default is `y ~ log(x)`}
}
\value{
figure of rarefaction curves
}
\description{
Rarefaction alpha index
}
\examples{
data(test_otu_data)
library(ggplot2)
prare <- ggrarecurve(test_otu_data,
               indexNames=c("Observe","Chao1","ACE"), 
               chunks=300) +
         theme(legend.spacing.y=unit(0.02,"cm"),
               legend.text=element_text(size=6))
}
\author{
Shuangbin Xu
}
