% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveSeq.R
\name{mapplyretrieveSeq}
\alias{mapplyretrieveSeq}
\title{Retriveing Sequencing from NCBI By mapply}
\usage{
mapplyretrieveSeq(idlist, files, databases = "protein", type = "fasta",
  times = 3, checkids = TRUE)
}
\arguments{
\item{files}{character, the file name specified by a double-quoted string.}

\item{databases}{character, the name of databases to use, default is `protein`,
if nucleotide sequences to retrieve set nuccore,see \code{\link[rentrez]{entrez_fetch}}.}

\item{type}{character, the format in which to get data,such as fasta, xml ...,
see \code{\link[rentrez]{entrez_fetch}}.}

\item{times}{integer, the time of sleeping, default is 3, 
meaning 3 seconds.}

\item{checkids}{bool, whether check the sequence of ids has been retrieved.
default is FALSE.}

\item{ids}{list, the accession number or accession.}
}
\description{
Retriveing sequences from NCBI with the accession ids.
}
\seealso{
[retrieveSeq]
}
\author{
ShuangbinXu
}
