% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-get-taxa.R
\name{mp_cal_abundance}
\alias{mp_cal_abundance}
\alias{mp_cal_abundance,MPSE-method}
\alias{mp_cal_abundance,MPSE}
\alias{mp_cal_abundance,tbl_mpse-method}
\alias{mp_cal_abundance,tbl_mpse}
\alias{mp_cal_abundance,grouped_df_mpse-method}
\alias{mp_cal_abundance,grouped_df_mpse}
\title{Calculate the (relative) abundance of each taxonomy class for each sample or group.}
\usage{
mp_cal_abundance(
  .data,
  .abundance = NULL,
  .group = NULL,
  relative = TRUE,
  action = "add",
  force = FALSE,
  ...
)

\S4method{mp_cal_abundance}{MPSE}(
  .data,
  .abundance = NULL,
  .group = NULL,
  relative = TRUE,
  action = "add",
  force = FALSE,
  ...
)

\S4method{mp_cal_abundance}{tbl_mpse}(
  .data,
  .abundance = NULL,
  .group = NULL,
  relative = TRUE,
  action = "add",
  force = FALSE,
  ...
)

\S4method{mp_cal_abundance}{grouped_df_mpse}(
  .data,
  .abundance = NULL,
  .group = NULL,
  relative = TRUE,
  action = "add",
  force = FALSE,
  ...
)
}
\arguments{
\item{.data}{MPSE or tbl_mpse object}

\item{.abundance}{the name of otu abundance to be calculated}

\item{.group}{the name of group to be calculated.}

\item{relative}{logical whether calculate the relative abundance.}

\item{action}{character, "add" joins the new information to the taxatree and 
otutree if they exists (default). In addition, All taxonomy class will be added 
the taxatree, and OTU (tip) information will be added to the otutree."only" return 
a non-redundant tibble with the just new information. "get" return 'taxatree' slot
which is a treedata object.}

\item{force}{logical whether calculate the relative abundance when the abundance 
is not be rarefied, default is FALSE.}

\item{...}{additional parameters.}
}
\value{
update object or tibble according the 'action'
}
\description{
Calculate the (relative) abundance of each taxonomy class for each sample or group.
}
