% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-get-dist.R
\name{mp_cal_dist}
\alias{mp_cal_dist}
\alias{mp_cal_dist,MPSE-method}
\alias{mp_cal_dist,MPSE}
\alias{mp_cal_dist,tbl_mpse-method}
\alias{mp_cal_dist,tbl_mpse}
\alias{mp_cal_dist,grouped_df_mpse-method}
\alias{mp_cal_dist,grouped_df_mpse}
\title{Calculate the distances between the samples with specified abundance.}
\usage{
mp_cal_dist(
  .data,
  .abundance,
  .env = NULL,
  distmethod = "bray",
  action = "add",
  scale = FALSE,
  ...
)

\S4method{mp_cal_dist}{MPSE}(
  .data,
  .abundance,
  .env = NULL,
  distmethod = "bray",
  action = "add",
  scale = FALSE,
  ...
)

\S4method{mp_cal_dist}{tbl_mpse}(
  .data,
  .abundance,
  .env = NULL,
  distmethod = "bray",
  action = "add",
  scale = FALSE,
  ...
)

\S4method{mp_cal_dist}{grouped_df_mpse}(
  .data,
  .abundance,
  .env = NULL,
  distmethod = "bray",
  action = "add",
  scale = FALSE,
  ...
)
}
\arguments{
\item{.data}{MPSE or tbl_mpse object}

\item{.abundance}{the name of otu abundance to be calculated}

\item{.env}{the column names of continuous environment factors,
default is NULL.}

\item{distmethod}{character the method to calculate distance.
option is "manhattan", "euclidean", "canberra", "bray", "kulczynski", 
"jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup",
"binomial", "chao", "cao" (implemented in vegdist of vegan), and
"w", "-1", "c", "wb", "r", "I", "e", "t", "me", "j", "sor", "m", "-2", "co"
"cc", "g", "-3", "l", "19", "hk", "rlb", "sim", "gl", "z" (implemented in 
betadiver of vegan), "maximum", "binary", "minkowski" (implemented in dist 
of stats), "unifrac", "weighted unifrac" (implement in phyloseq),}

\item{action}{character, "add" joins the distance data to the object, "only" return
a non-redundant tibble with the distance information. "get" return 'dist' object.}

\item{scale}{logical whether scale the metric of environment (.env is provided) before
the distance was calculated, default is FALSE. The environment matrix can be processed
when it was joined to the MPSE or tbl_mpse object.}

\item{...}{additional parameters.}
}
\value{
update object or tibble according the 'action'
}
\description{
Calculate the distances between the samples with specified abundance.
}
\examples{
data(mouse.time.mpse)
mouse.time.mpse \%<>\%
mp_decostand(.abundance=Abundance) \%>\% 
mp_cal_dist(.abundance=hellinger, distmethod="bray")
# Visualization
library(ggplot2)
tbl <- mouse.time.mpse \%>\% 
       mp_extract_sample 
tbl
tbl \%>\% 
  select(Sample, time, bray) \%>\% 
  unnest(cols=c(bray)) \%>\% 
  mutate(time=paste0(time, "-vs-", time[match(braySampley, Sample)])) \%>\% 
  dplyr::filter(time!="Late-vs-Early" & bray!=0) \%>\%
  ggplot(aes(x=time, y=bray)) + 
  geom_boxplot(aes(fill=time)) + 
  geom_jitter(width=0.1) + 
  xlab(NULL) + 
  theme(legend.position="none")
}
\author{
Shuangbin Xu
}
