% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-get-dist.R
\name{mp_cal_dist}
\alias{mp_cal_dist}
\alias{mp_cal_dist,MPSE-method}
\alias{mp_cal_dist,MPSE}
\alias{mp_cal_dist,tbl_mpse-method}
\alias{mp_cal_dist,tbl_mpse}
\alias{mp_cal_dist,grouped_df_mpse-method}
\alias{mp_cal_dist,grouped_df_mpse}
\title{Calculate the distances between the samples with specified abundance.}
\usage{
mp_cal_dist(
  .data,
  .abundance,
  .env = NULL,
  distmethod = "bray",
  action = "add",
  ...
)

\S4method{mp_cal_dist}{MPSE}(
  .data,
  .abundance,
  .env = NULL,
  distmethod = "bray",
  action = "add",
  ...
)

\S4method{mp_cal_dist}{tbl_mpse}(
  .data,
  .abundance,
  .env = NULL,
  distmethod = "bray",
  action = "add",
  ...
)

\S4method{mp_cal_dist}{grouped_df_mpse}(
  .data,
  .abundance,
  .env = NULL,
  distmethod = "bray",
  action = "add",
  ...
)
}
\arguments{
\item{.data}{MPSE or tbl_mpse object}

\item{.abundance}{the name of otu abundance to be calculated}

\item{.env}{the column names of continuous environment factors,
default is NULL.}

\item{distmethod}{character the method to calculate distance.
option is "manhattan", "euclidean", "canberra", "bray", "kulczynski", 
"jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup",
"binomial", "chao", "cao" (implemented in vegdist of vegan), and
"w", "-1", "c", "wb", "r", "I", "e", "t", "me", "j", "sor", "m", "-2", "co"
"cc", "g", "-3", "l", "19", "hk", "rlb", "sim", "gl", "z" (implemented in 
betadiver of vegan), "maximum", "binary", "minkowski" (implemented in dist 
of stats), "unifrac", "weighted unifrac" (implement in phyloseq),}

\item{action}{character, "add" joins the distance data to the object, "only" return
a non-redundant tibble with the distance information. "get" return 'dist' object.}

\item{...}{additional parameters.}
}
\value{
update object or tibble according the 'action'
}
\description{
Calculate the distances between the samples with specified abundance.
}
\examples{
data(mouse.time.mpse)
mouse.time.mpse \%>\%
mp_decostand(.abundance=Abundance) \%>\% 
mp_cal_dist(.abundance=hellinger)
}
\author{
Shuangbin Xu
}
