% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-mp_envfit.R
\name{mp_envfit}
\alias{mp_envfit}
\alias{mp_envfit,MPSE-method}
\alias{mp_envfit,MPSE}
\alias{mp_envfit,tbl_mpse-method}
\alias{mp_envfit,tbl_mpse}
\alias{mp_envfit,grouped_df_mpse-method}
\alias{mp_envfit,grouped_df_mpse}
\title{Fits an Environmental Vector or Factor onto an Ordination With MPSE or tbl_mpse Object}
\usage{
mp_envfit(
  .data,
  .ord,
  .env,
  .dim = 3,
  action = "add",
  permutations = 999,
  seed = 123,
  ...
)

\S4method{mp_envfit}{MPSE}(
  .data,
  .ord,
  .env,
  .dim = 3,
  action = "add",
  permutations = 999,
  seed = 123,
  ...
)

\S4method{mp_envfit}{tbl_mpse}(
  .data,
  .ord,
  .env,
  .dim = 3,
  action = "add",
  permutations = 999,
  seed = 123,
  ...
)

\S4method{mp_envfit}{grouped_df_mpse}(
  .data,
  .ord,
  .env,
  .dim = 3,
  action = "add",
  permutations = 999,
  seed = 123,
  ...
)
}
\arguments{
\item{.data}{MPSE or tbl_mpse object}

\item{.ord}{a name of ordination, option it is DCA, NMDS, RDA, CCA.}

\item{.env}{the names of columns of sample group or environment information.}

\item{.dim}{integer The number of dimensions to be returned, default is 3.}

\item{action}{character "add" joins the envfit result to internal attributes of the object, 
"only" return a non-redundant tibble with the envfit result. "get" return 'envfit' object can
be analyzed using the related vegan funtion.}

\item{permutations}{the number of permutations required, default is 999.}

\item{seed}{a random seed to make the analysis reproducible, default is 123.}

\item{...}{additional parameters see also 'vegan::envfit'}
}
\value{
update object according action
}
\description{
Fits an Environmental Vector or Factor onto an Ordination With MPSE or tbl_mpse Object
}
