% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-accessor.R
\docType{methods}
\name{taxonomy}
\alias{taxonomy}
\alias{taxonomy,MPSE-method}
\alias{taxonomy,MPSE}
\alias{taxonomy,tbl_mpse-method}
\alias{taxonomy,tbl_mpse}
\alias{taxonomy,grouped_df_mpse-method}
\alias{taxonomy,grouped_df_mpse}
\alias{mp_extract_taxonomy}
\alias{mp_extract_taxonomy,MPSE-method}
\alias{mp_extract_taxonomy,MPSE}
\alias{mp_extract_taxonomy,tbl_mpse-method}
\alias{mp_extract_taxonomy,tbl_mpse}
\alias{mp_extract_taxonomy,grouped_df_mpse-method}
\alias{mp_extract_taxonomy,grouped_df_mpse}
\title{extract the taxonomy annotation in MPSE object}
\usage{
taxonomy(x, ...)

\S4method{taxonomy}{MPSE}(x, ...)

\S4method{taxonomy}{tbl_mpse}(x, ...)

\S4method{taxonomy}{grouped_df_mpse}(x, ...)

mp_extract_taxonomy(x, ...)

\S4method{mp_extract_taxonomy}{MPSE}(x, ...)

\S4method{mp_extract_taxonomy}{tbl_mpse}(x, ...)

\S4method{mp_extract_taxonomy}{grouped_df_mpse}(x, ...)
}
\arguments{
\item{x}{MPSE object}

\item{...}{additional arguments}
}
\value{
data.frame contained taxonomy information

data.frame contained taxonomy annotation.
}
\description{
extract the taxonomy annotation in MPSE object
}
