% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-accessor.R
\docType{methods}
\name{mp_extract_tree}
\alias{mp_extract_tree}
\alias{mp_extract_tree,MPSE-method}
\alias{mp_extract_tree,MPSE}
\alias{mp_extract_tree,tbl_mpse-method}
\alias{mp_extract_tree,tbl_mpse}
\alias{mp_extract_tree,grouped_df_mpse-method}
\alias{mp_extract_tree,grouped_df_mpse}
\title{extract the taxonomy tree in MPSE object}
\usage{
mp_extract_tree(x, type = "taxatree", ...)

\S4method{mp_extract_tree}{MPSE}(x, type = "taxatree", ...)

\S4method{mp_extract_tree}{tbl_mpse}(x, type = "taxatree", ...)

\S4method{mp_extract_tree}{grouped_df_mpse}(x, type = "taxatree", ...)
}
\arguments{
\item{x}{MPSE object}

\item{type}{character taxatree or otutree}

\item{...}{additional arguments}
}
\value{
taxatree treedata object
}
\description{
extract the taxonomy tree in MPSE object
}
