% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-mp_diff_analysis.R
\name{mp_plot_diff_cladogram}
\alias{mp_plot_diff_cladogram}
\title{Visualizing the result of mp_diff_analysis with cladogram.}
\usage{
mp_plot_diff_cladogram(
  .data,
  .group,
  .size = "pvalue",
  taxa.class,
  removeUnknown = FALSE,
  layout = "radial",
  line.size = 0.12,
  label.size = 2.5,
  point.size = 1.5,
  tip.annot = TRUE,
  as.tiplab = TRUE,
  ...
)
}
\arguments{
\item{.data}{MPSE object or treedata which was from the taxatree slot 
after running the 'mp_diff_analysis'.}

\item{.group}{the column name for mapping the different color.}

\item{.size}{the column name for mapping the size of points, default is 'pvalue'.}

\item{taxa.class}{the taxonomy class name will be replaced shorthand, default is 
the one level above ‘OTU’.}

\item{removeUnknown}{logical, whether mask the unknown taxonomy information but differential species,
default is FALSE.}

\item{layout}{the layout of tree, default is 'radial', see also the 'layout' of 'ggtree'.}

\item{line.size}{the line size (width) of tree, default is 0.12 .}

\item{label.size}{the label size of differential taxa, default is 3.}

\item{point.size}{the point size of undifferential taxa (background tree), default is 1.5,}

\item{tip.annot}{logcial whether to replace the differential tip labels with shorthand,
default is TRUE.}

\item{as.tiplab}{logical whether to display the differential tip labels with 'geom_tiplab' 
of 'ggtree', default is TRUE, if it is FALSE, it will use 'geom_text_repel' of 'ggrepel'.}

\item{...}{additional parameters, meaningless now.}
}
\description{
Visualizing the result of mp_diff_analysis with cladogram.
}
\details{
The color scale of differential group can be designed by 'scale_fill_diff_cladogram'
}
