% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-mp_diff_analysis.R
\name{mp_plot_diff_res}
\alias{mp_plot_diff_res}
\alias{mp_plot_diff_res,MPSE-method}
\alias{mp_plot_diff_res,MPSE}
\alias{mp_plot_diff_res,tbl_mpse-method}
\alias{mp_plot_diff_res,tbl_mpse}
\alias{mp_plot_diff_res,grouped_df_mpse-method}
\alias{mp_plot_diff_res,grouped_df_mpse}
\title{The visualization of result of mp_diff_analysis}
\usage{
mp_plot_diff_res(.data, ...)

\S4method{mp_plot_diff_res}{MPSE}(.data, ...)

\S4method{mp_plot_diff_res}{tbl_mpse}(.data, ...)

\S4method{mp_plot_diff_res}{grouped_df_mpse}(.data, ...)
}
\arguments{
\item{.data}{MPSE or tbl_mpse after run mp_diff_analysis with \code{action="add"}}

\item{...}{additional parameters.}
}
\description{
The visualization of result of mp_diff_analysis
}
