% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-mp_rrarefy.R
\docType{methods}
\name{mp_rrarefy}
\alias{mp_rrarefy}
\alias{mp_rrarefy,matrix-method}
\alias{mp_rrarefy,matrix}
\alias{mp_rrarefy,data.frame-method}
\alias{mp_rrarefy,data.frame}
\alias{mp_rrarefy,phyloseq-method}
\alias{mp_rrarefy,phyloseq}
\alias{mp_rrarefy,tbl_ps-method}
\alias{mp_rrarefy,tbl_ps}
\alias{mp_rrarefy,grouped_df_ps-method}
\alias{mp_rrarefy,grouped_df_ps}
\title{mp_rrarefy method}
\source{
mp_rrarefy for matrix object is a wrapper method of vegan::rrarefy from the vegan 
package.
}
\usage{
mp_rrarefy(obj, raresize, trimOTUs = TRUE, seed = 123)

\S4method{mp_rrarefy}{matrix}(obj, raresize, trimOTUs = TRUE, seed = 123)

\S4method{mp_rrarefy}{data.frame}(obj, raresize, trimOTUs = TRUE, seed = 123)

\S4method{mp_rrarefy}{phyloseq}(obj, raresize, trimOTUs = TRUE, seed = 123)

\S4method{mp_rrarefy}{tbl_ps}(obj, raresize, trimOTUs = TRUE, seed = 123)

\S4method{mp_rrarefy}{grouped_df_ps}(obj, raresize, trimOTUs = TRUE, seed = 123)
}
\arguments{
\item{obj}{phyloseq or tbl_ps object}

\item{raresize}{integer Subsample size for rarefying community.}

\item{trimOTUs}{logical Whether to remove the otus that are no 
longer present in any sample after rarefaction}

\item{seed}{a random seed to make the rrarefy reproducible,
default is 123.}
}
\value{
update object
}
\description{
mp_rrarefy method
}
\seealso{
\link[vegan]{rrarefy}
}
