% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-mp_rrarefy.R
\docType{methods}
\name{mp_rrarefy}
\alias{mp_rrarefy}
\alias{mp_rrarefy,matrix-method}
\alias{mp_rrarefy,matrix}
\alias{mp_rrarefy,data.frame-method}
\alias{mp_rrarefy,data.frame}
\alias{mp_rrarefy,MPSE-method}
\alias{mp_rrarefy,MPSE}
\alias{mp_rrarefy,tbl_mpse-method}
\alias{mp_rrarefy,tbl_mpse}
\alias{mp_rrarefy,grouped_df_mpse-method}
\alias{mp_rrarefy,grouped_df_mpse}
\title{mp_rrarefy method}
\source{
mp_rrarefy for matrix object is a wrapper method of vegan::rrarefy from the vegan 
package.
}
\usage{
mp_rrarefy(.data, raresize, trimOTU = TRUE, seed = 123)

\S4method{mp_rrarefy}{matrix}(.data, raresize, trimOTU = TRUE, seed = 123)

\S4method{mp_rrarefy}{data.frame}(.data, raresize, trimOTU = TRUE, seed = 123)

\S4method{mp_rrarefy}{MPSE}(.data, raresize, trimOTU = TRUE, seed = 123)

\S4method{mp_rrarefy}{tbl_mpse}(.data, raresize, trimOTU = TRUE, seed = 123)

\S4method{mp_rrarefy}{grouped_df_mpse}(.data, raresize, trimOTU = TRUE, seed = 123)
}
\arguments{
\item{.data}{MPSE or tbl_mpse object}

\item{raresize}{integer Subsample size for rarefying community.}

\item{trimOTU}{logical Whether to remove the otus that are no 
longer present in any sample after rarefaction}

\item{seed}{a random seed to make the rrarefy reproducible,
default is 123.}
}
\value{
update object
}
\description{
mp_rrarefy method
}
\seealso{
\link[vegan]{rrarefy}
}
