% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-accessor.R
\name{MPSE-accessors}
\alias{MPSE-accessors}
\alias{[,MPSE,ANY,ANY,ANY-method}
\alias{colData<-,MPSE,DataFrame-method}
\alias{colData<-,MPSE,NULL-method}
\alias{otutree}
\alias{otutree,MPSE-method}
\alias{otutree,MPSE}
\alias{otutree<-}
\alias{otutree<-,MPSE,treedata-method}
\alias{otutree<-,MPSE}
\alias{otutree<-,MPSE,phylo-method}
\alias{otutree<-,MPSE,NULL-method}
\alias{taxatree}
\alias{taxatree,MPSE-method}
\alias{taxatree,MPSE}
\alias{taxatree<-}
\alias{taxatree<-,MPSE,treedata-method}
\alias{taxatree<-,MPSE}
\alias{taxatree<-,MPSE,NULL-method}
\alias{refsequence}
\alias{refsequence,MPSE-method}
\alias{refsequence,MPSE}
\alias{refsequence<-}
\alias{refsequence<-,MPSE,XStringSet-method}
\alias{refsequence<-,MPSE}
\alias{refsequence<-,MPSE,NULL-method}
\alias{rownames<-,MPSE-method}
\alias{rownames<-,MPSE}
\title{MPSE accessors}
\usage{
\S4method{[}{MPSE,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{colData}{MPSE,DataFrame}(x, ...) <- value

\S4method{colData}{MPSE,`NULL`}(x, ...) <- value

otutree(x, ...)

\S4method{otutree}{MPSE}(x, ...)

otutree(x, ...) <- value

\S4method{otutree}{MPSE,treedata}(x, ...) <- value

\S4method{otutree}{MPSE,phylo}(x, ...) <- value

\S4method{otutree}{MPSE,`NULL`}(x, ...) <- value

taxatree(x, ...)

\S4method{taxatree}{MPSE}(x, ...)

taxatree(x, ...) <- value

\S4method{taxatree}{MPSE,treedata}(x, ...) <- value

\S4method{taxatree}{MPSE,`NULL`}(x, ...) <- value

refsequence(x, ...)

\S4method{refsequence}{MPSE}(x, ...)

refsequence(x, ...) <- value

\S4method{refsequence}{MPSE,XStringSet}(x, ...) <- value

\S4method{refsequence}{MPSE,`NULL`}(x, ...) <- value

\S4method{rownames}{MPSE}(x) <- value
}
\arguments{
\item{x}{MPSE object}

\item{i, j, ...}{Indices specifying elements to extract or replace.
Indices are 'numeric' or 'character' vectors or empty (missing) or
NULL.  Numeric values are coerced to integer as by 'as.integer' 
(and hence truncated towards zero).  Character vectors will be matched 
to the 'names' of the object (or for matrices/arrays, the 'dimnames')}

\item{drop}{logical If 'TRUE' the result is coerced to the lowest 
possible dimension (see the examples).  This only works for extracting 
elements, not for the replacement.}

\item{value}{XStringSet object or NULL}
}
\description{
MPSE accessors
}
