% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-utilities.R
\name{multi.compare}
\alias{multi.compare}
\title{a container for performing two or more sample test.}
\usage{
multi.compare(fun = wilcox.test, data, feature, factorNames, ...)
}
\arguments{
\item{fun}{character, the method for test, optional ""}

\item{data}{data.frame, nrow sample * ncol feature+factorNames.}

\item{feature}{vector, the features wanted to test.}

\item{factorNames}{character, the name of a factor giving the corresponding groups.}

\item{..., }{additional arguments for fun.}
}
\value{
the result of fun, if fun is wilcox.test, it 
will return the list with class "htest".
}
\description{
a container for performing two or more sample test.
}
\examples{
datest <- data.frame(A=rnorm(1:10,mean=5),
                     B=rnorm(2:11, mean=6), 
                     group=c(rep("case",5),rep("control",5)))
head(datest)
multi.compare(fun=wilcox.test,data=datest,
              feature=c("A", "B"),factorNames="group")
}
\author{
Shuangbin Xu
}
