% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieveSeq.R
\name{retrieveSeq}
\alias{retrieveSeq}
\title{Retriveing Sequencing from NCBI}
\usage{
retrieveSeq(ids, files, databases = "protein", type = "fasta",
  times = 3, checkids = FALSE)
}
\arguments{
\item{ids}{vector, the accession number or accession.}

\item{files}{character, the file name specified by a double-quoted string.}

\item{databases}{character, the name of databases to use, default is `protein`,
if nucleotide sequences to retrieve set nuccore,
see also \code{\link[rentrez]{entrez_fetch}}.}

\item{type}{character, the format in which to get data,
such as fasta, xml ...,
see also \code{\link[rentrez]{entrez_fetch}}.}

\item{times}{integer, the time of sleeping, default is 3, 
meaning 3 seconds.}

\item{checkids}{logical, whether check the sequence of 
ids has been retrieved. default is FALSE.}
}
\value{
the files of sequences downloaded by ids
}
\description{
Retriveing sequences from NCBI with the accession ids.
}
\examples{
retrieveSeq(ids=c("ADM52729.1", "AAF82637.1"), 
            files="test.fasta",
            databases="protein",
            type="fasta",
            checkids=TRUE)
}
\author{
Shuangbin Xu
}
