% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml-method.R
\name{sampledflist}
\alias{sampledflist}
\title{Generate random data list from a original data.}
\usage{
sampledflist(dalist, bootnums = 30, ratio = 0.7,
  makerownames = FALSE, randomSeed = FALSE)
}
\arguments{
\item{dalist}{list, a list contained multi data.frame.}

\item{bootnums}{integer, the number of bootstrap iteration, default is 30.}

\item{ratio}{numeric, the ratios of each data.frame to keep.}

\item{makerownames}{logical, whether build row.names,default is FALSE.}

\item{randomSeed}{integer, A single integer value passed to ‘set.seed’, 
which is used to fix a seed for reproducibly random number, default is 1024.}
}
\value{
the list contained the data.frame generated by bootstrap iteration.
}
\description{
Generate random data list from a original data.
}
\examples{
data(iris)
irislist <- split(iris, iris$Species)
set.seed(1024)
irislist <- sampledflist(irislist)
}
\author{
Shuangbin Xu
}
