% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show}
\alias{show}
\title{show method}
\usage{
\S4method{show}{diffAnalysisClass}(object)
}
\arguments{
\item{objet}{object, `diffAnalysisClass` class}
}
\value{
print info
}
\description{
show method
}
\author{
Shuangbin Xu
}
