% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show,diffAnalysisClass-method}
\alias{show,diffAnalysisClass-method}
\title{method extensions to show for diffAnalysisClass objects.}
\usage{
\S4method{show}{diffAnalysisClass}(object)
}
\arguments{
\item{object}{object, `diffAnalysisClass` class}
}
\value{
print info
}
\description{
method extensions to show for diffAnalysisClass objects.
}
\examples{
data(kostic2012crc)
kostic2012crc
head(phyloseq::sample_data(kostic2012crc),3)
kostic2012crc <- phyloseq::rarefy_even_depth(kostic2012crc,rngseed=1024)
table(phyloseq::sample_data(kostic2012crc)$DIAGNOSIS)
diffres <- diffAnalysis(kostic2012crc, class="DIAGNOSIS",
                        mlfun="lda", filtermod="fdr",
                        firstcomfun = "kruskal.test",
                        firstalpha=0.05, strictmod=TRUE, 
                        secondcomfun = "wilcox.test",
                        submin=3, subclwilc=TRUE,
                        secondalpha=0.01, lda=3)
show(diffres)
}
\author{
Shuangbin Xu
}
